CIE D65 illuminant data

A dataset containing wavelengths at a 5 nm interval (300 nm to 830 nm) and the corresponding spectral energy irradiance normalized to 1 at 560 nm. Spectrum approximates the midday solar spectrum at middle latitude as 'corresponds' to the white point of a black body a 6504 K. Original data from http://files.cie.co.at/204.xls downloaded on 2014-07-25 The variables are as follows:

D65.illuminant.spct

Format

A source spectrum with 107 rows and 2 variables

Details

w.length (nm)

s.e.irrad (rel. units)

See also

Other Spectral data examples: A.illuminant.spct, black_body.spct, ccd.spct, clear.spct, clear_body.spct, filter_cps.mspct, green_leaf.spct, opaque.spct, photodiode.spct, polyester.spct, sun.daily.data, sun.daily.spct, sun.data, sun.spct, white_body.spct, white_led.cps_spct, white_led.raw_spct, white_led.source_spct, yellow_gel.spct

Examples

D65.illuminant.spct

#> Object: source_spct [107 x 2]
#> Wavelength range 300 to 830 nm, step 5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 560 nm 
#> 
#> # A tibble: 107 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000341
#>  2      305  0.016643
#>  3      310  0.032945
#>  4      315  0.117652
#>  5      320  0.202360
#>  6      325  0.286447
#>  7      330  0.370535
#>  8      335  0.385011
#>  9      340  0.399488
#> 10      345  0.424302
#> # ... with 97 more rows