Green Arabidopsis leaf spectral reflectance.

A dataset of total spectral reflectance expressed as fractions of one from the upper surface of a leaf of an Arabidopsis thaliana 'Ler' rosette.

Ler_leaf_rflt.spct

Format

An reflector_spct object with 1750 rows and 2 variables

Details

w.length (nm)

Rfr (0..1)

Note

Measured with a Jaz spectrometer from Ocean Optics (USA) configured with a PX Xenon lamp module and Spectroclip double integrating spheres.

See also

Other Spectral data examples: A.illuminant.spct, D65.illuminant.spct, Ler_leaf.spct, Ler_leaf_trns.spct, Ler_leaf_trns_i.spct, black_body.spct, ccd.spct, clear.spct, clear_body.spct, filter_cps.mspct, green_leaf.spct, opaque.spct, photodiode.spct, polyester.spct, sun.daily.data, sun.daily.spct, sun.data, sun.spct, white_body.spct, white_led.cps_spct, white_led.raw_spct, white_led.source_spct, yellow_gel.spct

Examples

Ler_leaf_rflt.spct

#> Object: reflector_spct [1,750 x 2]
#> Wavelength range 250 to 850 nm, step 0.07635631 to 0.3766982 nm 
#> 
#> # A tibble: 1,750 x 2
#>    w.length        Rfr
#>       <dbl>      <dbl>
#>  1 250.0000 0.04668778
#>  2 250.2261 0.04668481
#>  3 250.6028 0.04667985
#>  4 250.9794 0.04667489
#>  5 251.3561 0.04666993
#>  6 251.7327 0.04666497
#>  7 252.1092 0.04666001
#>  8 252.4858 0.04665505
#>  9 252.8623 0.04665010
#> 10 253.2387 0.04664514
#> # ... with 1,740 more rows