Convert transmittance into absorbance.

Function that converts transmittance (fraction) into absorbance (a.u.).

T2A(x, action, byref, clean, ...)

# S3 method for default
T2A(x, action = NULL, byref = FALSE, clean = TRUE, ...)

# S3 method for filter_spct
T2A(x, action = "add", byref = FALSE, clean = TRUE,
  ...)

# S3 method for filter_mspct
T2A(x, action = "add", byref = FALSE, clean = TRUE,
  ..., .parallel = FALSE, .paropts = NULL)

Arguments

x
an R object

action
character Allowed values "replace" and "add"

byref
logical indicating if new object will be created by reference or by copy of x

clean
logical replace off-boundary values before conversion

...
not used in current version

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

Methods (by class)

default: Default method for generic function

filter_spct: Method for filter spectra

filter_mspct: Method for collections of filter spectra

See also

Other quantity conversion functions: A2T, T2Afr, T2T, as_quantum, e2qmol_multipliers, e2quantum_multipliers, e2q, q2e