Convert transmittance into absorptance.

Function that coverts transmittance (fraction) into absorptance (fraction). If reflectance (fraction) is available, it allows conversions between internal and total absorptance.

T2Afr(x, action, byref, clean, ...)

# S3 method for default
T2Afr(x, action = NULL, byref = FALSE, clean = FALSE,
  ...)

# S3 method for filter_spct
T2Afr(x, action = "add", byref = FALSE,
  clean = FALSE, ...)

# S3 method for object_spct
T2Afr(x, action = "add", byref = FALSE,
  clean = FALSE, ...)

# S3 method for filter_mspct
T2Afr(x, action = "add", byref = FALSE,
  clean = FALSE, ...)

# S3 method for object_mspct
T2Afr(x, action = "add", byref = FALSE,
  clean = FALSE, ...)

Arguments

x
an R object

action
character Allowed values "replace" and "add"

byref
logical indicating if new object will be created by reference or by copy of x

clean
logical replace off-boundary values before conversion

...
not used in current version

Methods (by class)

default: Default method for generic function

filter_spct: Method for filter spectra

object_spct: Method for object spectra

filter_mspct: Method for collections of filter spectra

object_mspct: Method for collections of object spectra

See also

Other quantity conversion functions: A2T, T2A, T2T, as_energy, as_quantum_mol, as_quantum, e2qmol_multipliers, e2quantum_multipliers, e2q, q2e

Examples

T2Afr(Ler_leaf.spct)

#> Object: object_spct [2,401 x 4]
#> Wavelength range 250 to 850 nm, step 0.25 nm 
#> 
#> # A tibble: 2,401 x 4
#>    w.length   Tfr        Rfr   Afr
#>  *    <dbl> <dbl>      <dbl> <dbl>
#>  1   250.00     0 0.04668778     1
#>  2   250.25     0 0.04668449     1
#>  3   250.50     0 0.04668120     1
#>  4   250.75     0 0.04667791     1
#>  5   251.00     0 0.04667462     1
#>  6   251.25     0 0.04667132     1
#>  7   251.50     0 0.04666803     1
#>  8   251.75     0 0.04666474     1
#>  9   252.00     0 0.04666145     1
#> 10   252.25     0 0.04665816     1
#> # ... with 2,391 more rows