Collection-of-spectra copy-constructor

Return a copy of an R object with its class set to a given type of spectrum.

as.generic_mspct(x, force.spct.class = FALSE)

as.raw_mspct(x)

as.cps_mspct(x)

as.source_mspct(x, time.unit = c("second", "day", "exposure"),
  bswf.used = c("none", "unknown"),
  strict.range = getOption("photobiology.strict.range", default = FALSE))

as.response_mspct(x, time.unit = "second")

as.filter_mspct(x, Tfr.type = c("total", "internal"), strict.range = TRUE)

as.reflector_mspct(x, Rfr.type = c("total", "specular"),
  strict.range = TRUE)

as.object_mspct(x, Tfr.type = c("total", "internal"), Rfr.type = c("total",
  "specular"), strict.range = TRUE)

as.chroma_mspct(x)

Arguments

x
a list of spectral objects or a list of objects such as data frames that can be converted into spectral objects.

force.spct.class
logical indicating whether to change the class of members to generic_spct or retain the existing class.

time.unit
character A string, "second", "day" or "exposure"

bswf.used
character

strict.range
logical Flag indicating whether off-range values result in an error instead of a warning

Tfr.type
a character string, either "total" or "internal"

Rfr.type
a character string, either "total" or "specular"

Value

These functions return a copy of x converted into a given class of spectral collection object, if x is a valid argument to the corresponding set function.

Note

Members of generic_mspct objects can be heterogeneous: they can belong any class derived from generic_spct and class is not enforced. In this case when x is a list of data frames, force.spct.class = TRUE needs to be supplied.

See also

Other conversion of collections of spectra: join_mspct, split2mspct, subset2mspct