Coerce to a collection-of-spectra

Return a copy of an R object with its class set to a given type of spectrum.

as.object_mspct(x, ...)

# S3 method for default
as.object_mspct(x, ...)

# S3 method for data.frame
as.object_mspct(x, Tfr.type = c("total", "internal"),
  Rfr.type = c("total", "specular"), strict.range = TRUE, ...)

# S3 method for object_spct
as.object_mspct(x, ...)

# S3 method for list
as.object_mspct(x, Tfr.type = c("total", "internal"),
  Rfr.type = c("total", "specular"), strict.range = TRUE, ..., ncol = 1,
  byrow = FALSE)

Arguments

x
a list of spectral objects or a list of objects such as data frames that can be converted into spectral objects.

...
passed to individual spectrum object constructor

Tfr.type
a character string, either "total" or "internal"

Rfr.type
a character string, either "total" or "specular"

strict.range
logical Flag indicating how off-range values are handled

ncol
integer Number of 'virtual' columns in data

byrow
logical If ncol > 1 how to read in the data

Value

A copy of x converted into a object_mspct object.

Methods (by class)

default:

data.frame:

object_spct:

list:

See also