Coerce to a collection-of-spectra

Return a copy of an R object with its class set to a given type of spectrum.

as.response_mspct(x, ...)

# S3 method for default
as.response_mspct(x, ...)

# S3 method for data.frame
as.response_mspct(x, time.unit = "second", ...)

# S3 method for response_spct
as.response_mspct(x, ...)

# S3 method for list
as.response_mspct(x, time.unit = "second", ..., ncol = 1,
  byrow = FALSE)

# S3 method for matrix
as.response_mspct(x, w.length,
  spct.data.var = "s.e.response", multiplier = 1, byrow = NULL,
  spct.names = "spct_", ...)

Arguments

x
a list of spectral objects or a list of objects such as data frames that can be converted into spectral objects.

...
passed to individual spectrum object constructor

time.unit
character A string, "second", "day" or "exposure"

ncol
integer Number of 'virtual' columns in data

byrow
logical If ncol > 1 how to read in the data

w.length
numeric A vector of wavelengthvalues sorted in strictly ascending order (nm).

spct.data.var
character The name of the variable that will contain the spectral data. This indicates what physical quantity is stored in the matrix and the units of expression used.

multiplier
numeric A multiplier to be applied to the values in x to do unit or scale conversion.

spct.names
character Vector of names to be assigned to collection members, either of length 1, or with length equal to the number of spectra.

Value

A copy of x converted into a response_mspct object.

Methods (by class)

default:

data.frame:

response_spct:

list:

matrix:

Note

When x is a square matrix an explicit argument is needed for byrow to indicate how data in x should be read. In every case the length of the w.length vector must match one of the dimensions of x.

See also