Spectral weights

Calculate multipliers for selecting a range of wavelengths and optionally applying a biological spectral weighting function (BSWF) and wavelength normalization. This function returns numeric multipliers that can be used to select a waveband and apply a weight.

calc_multipliers(w.length, w.band, unit.out = "energy", unit.in = "energy",
  use.cached.mult = FALSE, fill = 0)

Arguments

w.length
numeric vector of wavelengths (nm).

w.band
waveband object.

unit.out
character A string: "photon" or "energy", default is "energy".

unit.in
character A string: "photon" or "energy", default is "energy".

use.cached.mult
logical Flag indicating whether multiplier values should be cached between calls.

fill
numeric If fill == NA then values returned for wavelengths outside the range of the waveband are set to NA.

Value

a numeric vector of multipliers of the same length as w.length.

See also

Other low-level functions operating on numeric vectors.: as_energy, as_quantum_mol, div_spectra, energy_irradiance, energy_ratio, insert_hinges, integrate_xy, interpolate_spectrum, irradiance, l_insert_hinges, oper_spectra, photon_irradiance, photon_ratio, photons_energy_ratio, prod_spectra, s_e_irrad2rgb, split_energy_irradiance, split_photon_irradiance, subt_spectra, sum_spectra, trim_tails, v_insert_hinges

Examples

with(sun.data, calc_multipliers(w.length, new_waveband(400,700),"photon"))

#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [6] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [11] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [16] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [21] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [26] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [31] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [41] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [51] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [56] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [61] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [66] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [71] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [76] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [81] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [86] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [91] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [101] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [106] 0.000000e+00 0.000000e+00 3.343690e-06 3.352049e-06 3.360408e-06
#> [111] 3.368768e-06 3.377127e-06 3.385486e-06 3.393845e-06 3.402204e-06
#> [116] 3.410564e-06 3.418923e-06 3.427282e-06 3.435641e-06 3.444001e-06
#> [121] 3.452360e-06 3.460719e-06 3.469078e-06 3.477437e-06 3.485797e-06
#> [126] 3.494156e-06 3.502515e-06 3.510874e-06 3.519234e-06 3.527593e-06
#> [131] 3.535952e-06 3.544311e-06 3.552670e-06 3.561030e-06 3.569389e-06
#> [136] 3.577748e-06 3.586107e-06 3.594467e-06 3.602826e-06 3.611185e-06
#> [141] 3.619544e-06 3.627904e-06 3.636263e-06 3.644622e-06 3.652981e-06
#> [146] 3.661340e-06 3.669700e-06 3.678059e-06 3.686418e-06 3.694777e-06
#> [151] 3.703137e-06 3.711496e-06 3.719855e-06 3.728214e-06 3.736573e-06
#> [156] 3.744933e-06 3.753292e-06 3.761651e-06 3.770010e-06 3.778370e-06
#> [161] 3.786729e-06 3.795088e-06 3.803447e-06 3.811806e-06 3.820166e-06
#> [166] 3.828525e-06 3.836884e-06 3.845243e-06 3.853603e-06 3.861962e-06
#> [171] 3.870321e-06 3.878680e-06 3.887039e-06 3.895399e-06 3.903758e-06
#> [176] 3.912117e-06 3.920476e-06 3.928836e-06 3.937195e-06 3.945554e-06
#> [181] 3.953913e-06 3.962273e-06 3.970632e-06 3.978991e-06 3.987350e-06
#> [186] 3.995709e-06 4.004069e-06 4.012428e-06 4.020787e-06 4.029146e-06
#> [191] 4.037506e-06 4.045865e-06 4.054224e-06 4.062583e-06 4.070942e-06
#> [196] 4.079302e-06 4.087661e-06 4.096020e-06 4.104379e-06 4.112739e-06
#> [201] 4.121098e-06 4.129457e-06 4.137816e-06 4.146175e-06 4.154535e-06
#> [206] 4.162894e-06 4.171253e-06 4.179612e-06 4.187972e-06 4.196331e-06
#> [211] 4.204690e-06 4.213049e-06 4.221408e-06 4.229768e-06 4.238127e-06
#> [216] 4.246486e-06 4.254845e-06 4.263205e-06 4.271564e-06 4.279923e-06
#> [221] 4.288282e-06 4.296641e-06 4.305001e-06 4.313360e-06 4.321719e-06
#> [226] 4.330078e-06 4.338438e-06 4.346797e-06 4.355156e-06 4.363515e-06
#> [231] 4.371875e-06 4.380234e-06 4.388593e-06 4.396952e-06 4.405311e-06
#> [236] 4.413671e-06 4.422030e-06 4.430389e-06 4.438748e-06 4.447108e-06
#> [241] 4.455467e-06 4.463826e-06 4.472185e-06 4.480544e-06 4.488904e-06
#> [246] 4.497263e-06 4.505622e-06 4.513981e-06 4.522341e-06 4.530700e-06
#> [251] 4.539059e-06 4.547418e-06 4.555777e-06 4.564137e-06 4.572496e-06
#> [256] 4.580855e-06 4.589214e-06 4.597574e-06 4.605933e-06 4.614292e-06
#> [261] 4.622651e-06 4.631010e-06 4.639370e-06 4.647729e-06 4.656088e-06
#> [266] 4.664447e-06 4.672807e-06 4.681166e-06 4.689525e-06 4.697884e-06
#> [271] 4.706244e-06 4.714603e-06 4.722962e-06 4.731321e-06 4.739680e-06
#> [276] 4.748040e-06 4.756399e-06 4.764758e-06 4.773117e-06 4.781477e-06
#> [281] 4.789836e-06 4.798195e-06 4.806554e-06 4.814913e-06 4.823273e-06
#> [286] 4.831632e-06 4.839991e-06 4.848350e-06 4.856710e-06 4.865069e-06
#> [291] 4.873428e-06 4.881787e-06 4.890146e-06 4.898506e-06 4.906865e-06
#> [296] 4.915224e-06 4.923583e-06 4.931943e-06 4.940302e-06 4.948661e-06
#> [301] 4.957020e-06 4.965379e-06 4.973739e-06 4.982098e-06 4.990457e-06
#> [306] 4.998816e-06 5.007176e-06 5.015535e-06 5.023894e-06 5.032253e-06
#> [311] 5.040612e-06 5.048972e-06 5.057331e-06 5.065690e-06 5.074049e-06
#> [316] 5.082409e-06 5.090768e-06 5.099127e-06 5.107486e-06 5.115846e-06
#> [321] 5.124205e-06 5.132564e-06 5.140923e-06 5.149282e-06 5.157642e-06
#> [326] 5.166001e-06 5.174360e-06 5.182719e-06 5.191079e-06 5.199438e-06
#> [331] 5.207797e-06 5.216156e-06 5.224515e-06 5.232875e-06 5.241234e-06
#> [336] 5.249593e-06 5.257952e-06 5.266312e-06 5.274671e-06 5.283030e-06
#> [341] 5.291389e-06 5.299748e-06 5.308108e-06 5.316467e-06 5.324826e-06
#> [346] 5.333185e-06 5.341545e-06 5.349904e-06 5.358263e-06 5.366622e-06
#> [351] 5.374981e-06 5.383341e-06 5.391700e-06 5.400059e-06 5.408418e-06
#> [356] 5.416778e-06 5.425137e-06 5.433496e-06 5.441855e-06 5.450215e-06
#> [361] 5.458574e-06 5.466933e-06 5.475292e-06 5.483651e-06 5.492011e-06
#> [366] 5.500370e-06 5.508729e-06 5.517088e-06 5.525448e-06 5.533807e-06
#> [371] 5.542166e-06 5.550525e-06 5.558884e-06 5.567244e-06 5.575603e-06
#> [376] 5.583962e-06 5.592321e-06 5.600681e-06 5.609040e-06 5.617399e-06
#> [381] 5.625758e-06 5.634117e-06 5.642477e-06 5.650836e-06 5.659195e-06
#> [386] 5.667554e-06 5.675914e-06 5.684273e-06 5.692632e-06 5.700991e-06
#> [391] 5.709350e-06 5.717710e-06 5.726069e-06 5.734428e-06 5.742787e-06
#> [396] 5.751147e-06 5.759506e-06 5.767865e-06 5.776224e-06 5.784583e-06
#> [401] 5.792943e-06 5.801302e-06 5.809661e-06 5.818020e-06 5.826380e-06
#> [406] 5.834739e-06 5.843098e-06 0.000000e+00 0.000000e+00 0.000000e+00
#> [411] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [416] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [421] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [426] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [431] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [436] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [441] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [446] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [451] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [456] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [461] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [466] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [471] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [476] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [481] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [486] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [491] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [496] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [501] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [506] 0.000000e+00 0.000000e+00 0.000000e+00

with(sun.data, calc_multipliers(w.length, new_waveband(400,700),"photon"), use.cached.mult = TRUE)

#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [6] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [11] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [16] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [21] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [26] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [31] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [41] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [51] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [56] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [61] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [66] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [71] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [76] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [81] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [86] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [91] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [101] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [106] 0.000000e+00 0.000000e+00 3.343690e-06 3.352049e-06 3.360408e-06
#> [111] 3.368768e-06 3.377127e-06 3.385486e-06 3.393845e-06 3.402204e-06
#> [116] 3.410564e-06 3.418923e-06 3.427282e-06 3.435641e-06 3.444001e-06
#> [121] 3.452360e-06 3.460719e-06 3.469078e-06 3.477437e-06 3.485797e-06
#> [126] 3.494156e-06 3.502515e-06 3.510874e-06 3.519234e-06 3.527593e-06
#> [131] 3.535952e-06 3.544311e-06 3.552670e-06 3.561030e-06 3.569389e-06
#> [136] 3.577748e-06 3.586107e-06 3.594467e-06 3.602826e-06 3.611185e-06
#> [141] 3.619544e-06 3.627904e-06 3.636263e-06 3.644622e-06 3.652981e-06
#> [146] 3.661340e-06 3.669700e-06 3.678059e-06 3.686418e-06 3.694777e-06
#> [151] 3.703137e-06 3.711496e-06 3.719855e-06 3.728214e-06 3.736573e-06
#> [156] 3.744933e-06 3.753292e-06 3.761651e-06 3.770010e-06 3.778370e-06
#> [161] 3.786729e-06 3.795088e-06 3.803447e-06 3.811806e-06 3.820166e-06
#> [166] 3.828525e-06 3.836884e-06 3.845243e-06 3.853603e-06 3.861962e-06
#> [171] 3.870321e-06 3.878680e-06 3.887039e-06 3.895399e-06 3.903758e-06
#> [176] 3.912117e-06 3.920476e-06 3.928836e-06 3.937195e-06 3.945554e-06
#> [181] 3.953913e-06 3.962273e-06 3.970632e-06 3.978991e-06 3.987350e-06
#> [186] 3.995709e-06 4.004069e-06 4.012428e-06 4.020787e-06 4.029146e-06
#> [191] 4.037506e-06 4.045865e-06 4.054224e-06 4.062583e-06 4.070942e-06
#> [196] 4.079302e-06 4.087661e-06 4.096020e-06 4.104379e-06 4.112739e-06
#> [201] 4.121098e-06 4.129457e-06 4.137816e-06 4.146175e-06 4.154535e-06
#> [206] 4.162894e-06 4.171253e-06 4.179612e-06 4.187972e-06 4.196331e-06
#> [211] 4.204690e-06 4.213049e-06 4.221408e-06 4.229768e-06 4.238127e-06
#> [216] 4.246486e-06 4.254845e-06 4.263205e-06 4.271564e-06 4.279923e-06
#> [221] 4.288282e-06 4.296641e-06 4.305001e-06 4.313360e-06 4.321719e-06
#> [226] 4.330078e-06 4.338438e-06 4.346797e-06 4.355156e-06 4.363515e-06
#> [231] 4.371875e-06 4.380234e-06 4.388593e-06 4.396952e-06 4.405311e-06
#> [236] 4.413671e-06 4.422030e-06 4.430389e-06 4.438748e-06 4.447108e-06
#> [241] 4.455467e-06 4.463826e-06 4.472185e-06 4.480544e-06 4.488904e-06
#> [246] 4.497263e-06 4.505622e-06 4.513981e-06 4.522341e-06 4.530700e-06
#> [251] 4.539059e-06 4.547418e-06 4.555777e-06 4.564137e-06 4.572496e-06
#> [256] 4.580855e-06 4.589214e-06 4.597574e-06 4.605933e-06 4.614292e-06
#> [261] 4.622651e-06 4.631010e-06 4.639370e-06 4.647729e-06 4.656088e-06
#> [266] 4.664447e-06 4.672807e-06 4.681166e-06 4.689525e-06 4.697884e-06
#> [271] 4.706244e-06 4.714603e-06 4.722962e-06 4.731321e-06 4.739680e-06
#> [276] 4.748040e-06 4.756399e-06 4.764758e-06 4.773117e-06 4.781477e-06
#> [281] 4.789836e-06 4.798195e-06 4.806554e-06 4.814913e-06 4.823273e-06
#> [286] 4.831632e-06 4.839991e-06 4.848350e-06 4.856710e-06 4.865069e-06
#> [291] 4.873428e-06 4.881787e-06 4.890146e-06 4.898506e-06 4.906865e-06
#> [296] 4.915224e-06 4.923583e-06 4.931943e-06 4.940302e-06 4.948661e-06
#> [301] 4.957020e-06 4.965379e-06 4.973739e-06 4.982098e-06 4.990457e-06
#> [306] 4.998816e-06 5.007176e-06 5.015535e-06 5.023894e-06 5.032253e-06
#> [311] 5.040612e-06 5.048972e-06 5.057331e-06 5.065690e-06 5.074049e-06
#> [316] 5.082409e-06 5.090768e-06 5.099127e-06 5.107486e-06 5.115846e-06
#> [321] 5.124205e-06 5.132564e-06 5.140923e-06 5.149282e-06 5.157642e-06
#> [326] 5.166001e-06 5.174360e-06 5.182719e-06 5.191079e-06 5.199438e-06
#> [331] 5.207797e-06 5.216156e-06 5.224515e-06 5.232875e-06 5.241234e-06
#> [336] 5.249593e-06 5.257952e-06 5.266312e-06 5.274671e-06 5.283030e-06
#> [341] 5.291389e-06 5.299748e-06 5.308108e-06 5.316467e-06 5.324826e-06
#> [346] 5.333185e-06 5.341545e-06 5.349904e-06 5.358263e-06 5.366622e-06
#> [351] 5.374981e-06 5.383341e-06 5.391700e-06 5.400059e-06 5.408418e-06
#> [356] 5.416778e-06 5.425137e-06 5.433496e-06 5.441855e-06 5.450215e-06
#> [361] 5.458574e-06 5.466933e-06 5.475292e-06 5.483651e-06 5.492011e-06
#> [366] 5.500370e-06 5.508729e-06 5.517088e-06 5.525448e-06 5.533807e-06
#> [371] 5.542166e-06 5.550525e-06 5.558884e-06 5.567244e-06 5.575603e-06
#> [376] 5.583962e-06 5.592321e-06 5.600681e-06 5.609040e-06 5.617399e-06
#> [381] 5.625758e-06 5.634117e-06 5.642477e-06 5.650836e-06 5.659195e-06
#> [386] 5.667554e-06 5.675914e-06 5.684273e-06 5.692632e-06 5.700991e-06
#> [391] 5.709350e-06 5.717710e-06 5.726069e-06 5.734428e-06 5.742787e-06
#> [396] 5.751147e-06 5.759506e-06 5.767865e-06 5.776224e-06 5.784583e-06
#> [401] 5.792943e-06 5.801302e-06 5.809661e-06 5.818020e-06 5.826380e-06
#> [406] 5.834739e-06 5.843098e-06 0.000000e+00 0.000000e+00 0.000000e+00
#> [411] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [416] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [421] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [426] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [431] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [436] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [441] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [446] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [451] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [456] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [461] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [466] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [471] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [476] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [481] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [486] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [491] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [496] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [501] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [506] 0.000000e+00 0.000000e+00 0.000000e+00