Light-source spectral output

Calculate interpolated values by interpolation from user-supplied spectral emission data or by name for light source data included in the packages photobiologySun, photobiologyLamps, or photobiologyLEDs, scaling the values.

calc_source_output(w.length.out, w.length.in, s.irrad.in, unit.in = "energy",
  scaled = NULL, fill = NA)

Arguments

w.length.out
numeric vector of wavelengths (nm) for output

w.length.in
numeric vector of wavelengths (nm) for input

s.irrad.in
numeric vector of spectral transmittance value (fractions or percent)

unit.in
a character string "energy" or "photon"

scaled
NULL, "peak", "area"; div ignored if !is.null(scaled)

fill
if NA, no extrapolation is done, and NA is returned for wavelengths outside the range of the input. If NULL then the tails are deleted. If 0 then the tails are set to zero.

Value

a source_spct with three numeric vectors with wavelength values (w.length), scaled and interpolated spectral energy irradiance (s.e.irrad), scaled and interpolated spectral photon irradiance values (s.q.irrad).

Note

This is a convenience function that adds no new functionality but makes it a little easier to plot lamp spectral emission data consistently. It automates interpolation, extrapolation/trimming and scaling.

Examples

with(sun.data, calc_source_output(290:1100, w.length.in=w.length, s.irrad.in=s.e.irrad))

#> Object: source_spct [811 x 3]
#> Wavelength range 290 to 1100 nm, step 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 811 x 3
#>    w.length    s.e.irrad    s.q.irrad
#>       <int>        <dbl>        <dbl>
#>  1      290           NA           NA
#>  2      291           NA           NA
#>  3      292           NA           NA
#>  4      293 2.609665e-06 6.391730e-12
#>  5      294 6.142401e-06 1.509564e-11
#>  6      295 2.176175e-05 5.366385e-11
#>  7      296 6.780119e-05 1.677626e-10
#>  8      297 1.533491e-04 3.807181e-10
#>  9      298 3.669677e-04 9.141345e-10
#> 10      299 7.845430e-04 1.960893e-09
#> # ... with 801 more rows