Check validity of spectral objects

Check that an R object contains the expected data members.

check_spct(x, byref, strict.range, ...)

# S3 method for default
check_spct(x, byref = FALSE, strict.range = NA, ...)

# S3 method for generic_spct
check_spct(x, byref = TRUE, strict.range = NA,
  multiple.wl = getMultipleWl(x), ...)

# S3 method for raw_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for cps_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for filter_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for reflector_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for object_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for response_spct
check_spct(x, byref = TRUE, strict.range = NA,
  multiple.wl = getMultipleWl(x), ...)

# S3 method for source_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

# S3 method for chroma_spct
check_spct(x, byref = TRUE,
  strict.range = getOption("photobiology.strict.range", default = FALSE),
  multiple.wl = getMultipleWl(x), ...)

Arguments

x
An R object

byref
logical indicating if new object will be created by reference or by copy of x

strict.range
logical indicating whether off-range values result in an error instead of a warning, NA disables the test.

...
additional param possible in derived methods

multiple.wl
numeric Maximum number of repeated w.length entries with same value.

Methods (by class)

default: Default for generic function.

generic_spct: Specialization for generic_spct.

raw_spct: Specialization for cps_spct.

cps_spct: Specialization for cps_spct.

filter_spct: Specialization for filter_spct.

reflector_spct: Specialization for reflector_spct.

object_spct: Specialization for object_spct.

response_spct: Specialization for response_spct.

source_spct: Specialization for source_spct.

chroma_spct: Specialization for chroma_spct.

See also

Other data validity check functions: check_spectrum, check_w.length

Examples

check_spct(sun.spct)

#> Object: source_spct [522 x 3]
#> Wavelength range 280 to 800 nm, step 0.9230769 to 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 522 x 3
#>    w.length s.e.irrad s.q.irrad
#>       <dbl>     <dbl>     <dbl>
#>  1 280.0000         0         0
#>  2 280.9231         0         0
#>  3 281.8462         0         0
#>  4 282.7692         0         0
#>  5 283.6923         0         0
#>  6 284.6154         0         0
#>  7 285.5385         0         0
#>  8 286.4615         0         0
#>  9 287.3846         0         0
#> 10 288.3077         0         0
#> # ... with 512 more rows