Clean (=replace) off-range values in a spectrum

These functions implement the equivalent of replace() but for spectral objects instead of vectors.

clean(x, range, range.s.data, fill, ...)

# S3 method for default
clean(x, range, range.s.data, fill, ...)

# S3 method for source_spct
clean(x, range = x, range.s.data = c(0, NA),
  fill = range.s.data, unit.out = getOption("photobiology.radiation.unit",
  default = "energy"), ...)

# S3 method for filter_spct
clean(x, range = x, range.s.data = NULL,
  fill = range.s.data, qty.out = getOption("photobiology.filter.qty",
  default = "transmittance"), ...)

# S3 method for reflector_spct
clean(x, range = x, range.s.data = c(0, 1),
  fill = range.s.data, ...)

# S3 method for response_spct
clean(x, range = x, range.s.data = c(0, NA),
  fill = range.s.data, unit.out = getOption("photobiology.radiation.unit",
  default = "energy"), ...)

# S3 method for cps_spct
clean(x, range = x, range.s.data = c(0, NA),
  fill = range.s.data, ...)

# S3 method for raw_spct
clean(x, range = x, range.s.data = c(NA_real_, NA_real_),
  fill = range.s.data, ...)

# S3 method for generic_spct
clean(x, range = x, range.s.data = c(NA_real_,
  NA_real_), fill = range.s.data, col.names, ...)

# S3 method for source_mspct
clean(x, range = NULL, range.s.data = c(0, NA),
  fill = range.s.data, unit.out = getOption("photobiology.radiation.unit",
  default = "energy"), ..., .parallel = FALSE, .paropts = NULL)

# S3 method for filter_mspct
clean(x, range = NULL, range.s.data = NULL,
  fill = range.s.data, qty.out = getOption("photobiology.filter.qty",
  default = "transmittance"), ..., .parallel = FALSE, .paropts = NULL)

# S3 method for reflector_mspct
clean(x, range = NULL, range.s.data = c(0, 1),
  fill = range.s.data, ..., .parallel = FALSE, .paropts = NULL)

# S3 method for response_mspct
clean(x, range = NULL, range.s.data = c(0, NA),
  fill = range.s.data, unit.out = getOption("photobiology.radiation.unit",
  default = "energy"), ..., .parallel = FALSE, .paropts = NULL)

# S3 method for cps_mspct
clean(x, range = NULL, range.s.data = c(0, NA),
  fill = range.s.data, ..., .parallel = FALSE, .paropts = NULL)

# S3 method for raw_mspct
clean(x, range = NULL, range.s.data = c(0, NA),
  fill = range.s.data, ..., .parallel = FALSE, .paropts = NULL)

# S3 method for generic_mspct
clean(x, range = x, range.s.data = c(NA_real_,
  NA_real_), fill = range.s.data, col.names, ..., .parallel = FALSE,
  .paropts = NULL)

Arguments

x
an R object

range
numeric vector of wavelengths

range.s.data
numeric vector of length two giving the allowable range for the spectral data.

fill
numeric vector of length 1 or 2, giving the replacement values to use at each extreme of the range.

...
currently ignored

unit.out
character string with allowed values "energy", and "photon", or its alias "quantum"

qty.out
character string with allowed values "energy", and "photon", or its alias "quantum"

col.names
character The name of the variable to clean

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

Value

A copy of x, possibly with some of the spectral data values replaced by the value passed to fill.

Methods (by class)

default: Default for generic function

source_spct: Replace off-range values in a source spectrum

filter_spct: Replace off-range values in a filter spectrum

reflector_spct: Replace off-range values in a reflector spectrum

response_spct: Replace off-range values in a response spectrum

cps_spct: Replace off-range values in a counts per second spectrum

raw_spct: Replace off-range values in a raw counts spectrum

generic_spct: Replace off-range values in a generic spectrum

source_mspct:

filter_mspct:

reflector_mspct:

response_mspct:

cps_mspct:

raw_mspct:

generic_mspct: