Clip head and/or tail of a spectrum

Clipping of head and tail of a spectrum based on wavelength limits, no interpolation used.

clip_wl(x, range, ...)

# S3 method for default
clip_wl(x, range, ...)

# S3 method for generic_spct
clip_wl(x, range = NULL, ...)

# S3 method for generic_mspct
clip_wl(x, range = NULL, ...)

# S3 method for waveband
clip_wl(x, range = NULL, ...)

# S3 method for list
clip_wl(x, range = NULL, ...)

Arguments

x
an R object

range
a numeric vector of length two, or any other object for which function range() will return range of walengths expressed in nanometres.

...
not used

Value

an R object of same class as input, most frequently of a shorter length, and never longer.

Methods (by class)

default: Default for generic function

generic_spct: Clip an object of class "generic_spct" or derived.

generic_mspct: Clip an object of class "generic_mspct" or derived.

waveband: Clip an object of class "waveband".

list: Clip a list (of objects of class "waveband").

Note

The condition tested is wl >= range[1] & wl < (range[2] + 1e-13).

See also

Other trim functions: trim_spct, trim_tails, trim_waveband, trim_wl

Examples

clip_wl(sun.spct, range = c(400, 500))

#> Object: source_spct [101 x 3]
#> Wavelength range 400 to 500 nm, step 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 101 × 3
#>    w.length s.e.irrad    s.q.irrad
#> *     <dbl>     <dbl>        <dbl>
#> 1       400 0.6081049 2.033314e-06
#> 2       401 0.6261742 2.098967e-06
#> 3       402 0.6497388 2.183388e-06
#> 4       403 0.6207287 2.091091e-06
#> 5       404 0.6370489 2.151395e-06
#> 6       405 0.6263786 2.120596e-06
#> 7       406 0.5996440 2.035099e-06
#> 8       407 0.5951226 2.024729e-06
#> 9       408 0.6256209 2.133720e-06
#> 10      409 0.6615421 2.261761e-06
#> # ... with 91 more rows

clip_wl(sun.spct, range = c(NA, 500))

#> Object: source_spct [222 x 3]
#> Wavelength range 280 to 500 nm, step 0.9230769 to 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 222 × 3
#>    w.length s.e.irrad s.q.irrad
#> *     <dbl>     <dbl>     <dbl>
#> 1  280.0000         0         0
#> 2  280.9231         0         0
#> 3  281.8462         0         0
#> 4  282.7692         0         0
#> 5  283.6923         0         0
#> 6  284.6154         0         0
#> 7  285.5385         0         0
#> 8  286.4615         0         0
#> 9  287.3846         0         0
#> 10 288.3077         0         0
#> # ... with 212 more rows

clip_wl(sun.spct, range = c(400, NA))

#> Object: source_spct [401 x 3]
#> Wavelength range 400 to 800 nm, step 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 401 × 3
#>    w.length s.e.irrad    s.q.irrad
#> *     <dbl>     <dbl>        <dbl>
#> 1       400 0.6081049 2.033314e-06
#> 2       401 0.6261742 2.098967e-06
#> 3       402 0.6497388 2.183388e-06
#> 4       403 0.6207287 2.091091e-06
#> 5       404 0.6370489 2.151395e-06
#> 6       405 0.6263786 2.120596e-06
#> 7       406 0.5996440 2.035099e-06
#> 8       407 0.5951226 2.024729e-06
#> 9       408 0.6256209 2.133720e-06
#> 10      409 0.6615421 2.261761e-06
#> # ... with 391 more rows