Color of an object

Equivalent RGB color of an object such as a spectrum, wavelength or waveband.

color_of(x, ...)

# S3 method for default
color_of(x, ...)

# S3 method for numeric
color_of(x, type = "CMF", ...)

# S3 method for list
color_of(x, short.names = TRUE, type = "CMF", ...)

# S3 method for waveband
color_of(x, short.names = TRUE, type = "CMF", ...)

# S3 method for source_spct
color_of(x, type = "CMF", ...)

# S3 method for source_mspct
color_of(x, ..., idx = !is.null(names(x)))

colour_of(x, ...)

color(x, ...)

Arguments

x
an R object.

...
ignored (possibly used by derived methods).

type
character telling whether "CMF", "CC", or "both" should be returned.

short.names
logical indicating whether to use short or long names for wavebands

idx
logical whether to add a column with the names of the elements of spct

Value

A color definition in hexadecimal format as a character string of 7 characters, "#" followed by the red, blue, and green values in hexadecimal (scaled to 0 ... 255). In the case of the specialization for list, a list of such definitions is returned. In the case of a collection of spectra, a data.frame with one column with such definitions and by default an additional column with names of the spectra as index.

Methods (by class)

default: Default method (returns always "black").

numeric: Method that returns Color definitions corresponding to numeric values representing a wavelengths in nm.

list: Method that returns Color of elements in a list.

waveband: Color at midpoint of a waveband object.

source_spct:

source_mspct:

Note

When x is a list but not a waveband, if a method color_of is not available for the class of each element of the list, then color_of.default will be called.

Deprecated

Use of color() is deprecated as this wrapper function may be removed in future versions of the package because of name clashes. Use color_of() instead.

Examples

wavelengths <- c(300, 420, 500, 600, NA) # nanometres
color_of(wavelengths)

#> [1] "#000000" "#1B00FF" "#00A92E" "#FF2E00" NA       

color_of(waveband(c(300,400)))

#> range.300.400.CMF 
#>         "#02000F" 

color_of(list(blue = waveband(c(400,480)), red = waveband(c(600,700))))

#>  blue.CMF   red.CMF 
#> "#0000FF" "#FF0000" 

color_of(numeric())

#> character(0)

color_of(NA_real_)

#> [1] NA

color_of(sun.spct)

#> source CMF 
#>  "#544F4B"