Convert energy-based quantities into photon-based quantities.

Function that coverts spectral energy irradiance into spectral photon irradiance (molar).

e2q(x, action, byref, ...)

# S3 method for default
e2q(x, action = "add", byref = FALSE, ...)

# S3 method for source_spct
e2q(x, action = "add", byref = FALSE, ...)

# S3 method for response_spct
e2q(x, action = "add", byref = FALSE, ...)

# S3 method for source_mspct
e2q(x, action = "add", byref = FALSE, ...)

# S3 method for response_mspct
e2q(x, action = "add", byref = FALSE, ...)

Arguments

x
an R object

action
a character string

byref
logical indicating if new object will be created by reference or by copy of x

...
not used in current version

Methods (by class)

default: Default method

source_spct: Method for spectral irradiance

response_spct: Method for spectral responsiveness

source_mspct: Method for collections of (light) source spectra

response_mspct: Method for for collections of response spectra

See also

Other quantity conversion functions: A2T, T2Afr, T2A, T2T, as_energy, as_quantum_mol, as_quantum, e2qmol_multipliers, e2quantum_multipliers, q2e