Calculate energy to quantum multipliers

Gives multipliers as a function of wavelength, for converting from energy to photon (quantum) units (number of photons as default, or moles of photons).

e2quantum_multipliers(w.length, molar = FALSE)

Arguments

w.length
numeric Vector of wavelengths (nm)

molar
logical Flag indicating whether output should be in moles or numbers

Value

A numeric vector of multipliers

See also

Other quantity conversion functions: A2T, T2Afr, T2A, T2T, as_quantum, e2qmol_multipliers, e2q, q2e

Examples

with(sun.data, e2quantum_multipliers(w.length))

#>   [1] 1.474975e+18 1.480009e+18 1.485043e+18 1.490077e+18 1.495111e+18
#>   [6] 1.500145e+18 1.505179e+18 1.510213e+18 1.515247e+18 1.520281e+18
#>  [11] 1.525315e+18 1.530349e+18 1.535383e+18 1.540417e+18 1.545451e+18
#>  [16] 1.550485e+18 1.555519e+18 1.560553e+18 1.565587e+18 1.570621e+18
#>  [21] 1.575656e+18 1.580690e+18 1.585724e+18 1.590758e+18 1.595792e+18
#>  [26] 1.600826e+18 1.605860e+18 1.610894e+18 1.615928e+18 1.620962e+18
#>  [31] 1.625996e+18 1.631030e+18 1.636064e+18 1.641098e+18 1.646132e+18
#>  [36] 1.651166e+18 1.656200e+18 1.661234e+18 1.666268e+18 1.671302e+18
#>  [41] 1.676336e+18 1.681370e+18 1.686404e+18 1.691439e+18 1.696473e+18
#>  [46] 1.701507e+18 1.706541e+18 1.711575e+18 1.716609e+18 1.721643e+18
#>  [51] 1.726677e+18 1.731711e+18 1.736745e+18 1.741779e+18 1.746813e+18
#>  [56] 1.751847e+18 1.756881e+18 1.761915e+18 1.766949e+18 1.771983e+18
#>  [61] 1.777017e+18 1.782051e+18 1.787085e+18 1.792119e+18 1.797153e+18
#>  [66] 1.802187e+18 1.807222e+18 1.812256e+18 1.817290e+18 1.822324e+18
#>  [71] 1.827358e+18 1.832392e+18 1.837426e+18 1.842460e+18 1.847494e+18
#>  [76] 1.852528e+18 1.857562e+18 1.862596e+18 1.867630e+18 1.872664e+18
#>  [81] 1.877698e+18 1.882732e+18 1.887766e+18 1.892800e+18 1.897834e+18
#>  [86] 1.902868e+18 1.907902e+18 1.912936e+18 1.917970e+18 1.923005e+18
#>  [91] 1.928039e+18 1.933073e+18 1.938107e+18 1.943141e+18 1.948175e+18
#>  [96] 1.953209e+18 1.958243e+18 1.963277e+18 1.968311e+18 1.973345e+18
#> [101] 1.978379e+18 1.983413e+18 1.988447e+18 1.993481e+18 1.998515e+18
#> [106] 2.003549e+18 2.008583e+18 2.013617e+18 2.018651e+18 2.023685e+18
#> [111] 2.028719e+18 2.033753e+18 2.038788e+18 2.043822e+18 2.048856e+18
#> [116] 2.053890e+18 2.058924e+18 2.063958e+18 2.068992e+18 2.074026e+18
#> [121] 2.079060e+18 2.084094e+18 2.089128e+18 2.094162e+18 2.099196e+18
#> [126] 2.104230e+18 2.109264e+18 2.114298e+18 2.119332e+18 2.124366e+18
#> [131] 2.129400e+18 2.134434e+18 2.139468e+18 2.144502e+18 2.149536e+18
#> [136] 2.154570e+18 2.159605e+18 2.164639e+18 2.169673e+18 2.174707e+18
#> [141] 2.179741e+18 2.184775e+18 2.189809e+18 2.194843e+18 2.199877e+18
#> [146] 2.204911e+18 2.209945e+18 2.214979e+18 2.220013e+18 2.225047e+18
#> [151] 2.230081e+18 2.235115e+18 2.240149e+18 2.245183e+18 2.250217e+18
#> [156] 2.255251e+18 2.260285e+18 2.265319e+18 2.270353e+18 2.275388e+18
#> [161] 2.280422e+18 2.285456e+18 2.290490e+18 2.295524e+18 2.300558e+18
#> [166] 2.305592e+18 2.310626e+18 2.315660e+18 2.320694e+18 2.325728e+18
#> [171] 2.330762e+18 2.335796e+18 2.340830e+18 2.345864e+18 2.350898e+18
#> [176] 2.355932e+18 2.360966e+18 2.366000e+18 2.371034e+18 2.376068e+18
#> [181] 2.381102e+18 2.386136e+18 2.391171e+18 2.396205e+18 2.401239e+18
#> [186] 2.406273e+18 2.411307e+18 2.416341e+18 2.421375e+18 2.426409e+18
#> [191] 2.431443e+18 2.436477e+18 2.441511e+18 2.446545e+18 2.451579e+18
#> [196] 2.456613e+18 2.461647e+18 2.466681e+18 2.471715e+18 2.476749e+18
#> [201] 2.481783e+18 2.486817e+18 2.491851e+18 2.496885e+18 2.501919e+18
#> [206] 2.506954e+18 2.511988e+18 2.517022e+18 2.522056e+18 2.527090e+18
#> [211] 2.532124e+18 2.537158e+18 2.542192e+18 2.547226e+18 2.552260e+18
#> [216] 2.557294e+18 2.562328e+18 2.567362e+18 2.572396e+18 2.577430e+18
#> [221] 2.582464e+18 2.587498e+18 2.592532e+18 2.597566e+18 2.602600e+18
#> [226] 2.607634e+18 2.612668e+18 2.617702e+18 2.622737e+18 2.627771e+18
#> [231] 2.632805e+18 2.637839e+18 2.642873e+18 2.647907e+18 2.652941e+18
#> [236] 2.657975e+18 2.663009e+18 2.668043e+18 2.673077e+18 2.678111e+18
#> [241] 2.683145e+18 2.688179e+18 2.693213e+18 2.698247e+18 2.703281e+18
#> [246] 2.708315e+18 2.713349e+18 2.718383e+18 2.723417e+18 2.728451e+18
#> [251] 2.733485e+18 2.738520e+18 2.743554e+18 2.748588e+18 2.753622e+18
#> [256] 2.758656e+18 2.763690e+18 2.768724e+18 2.773758e+18 2.778792e+18
#> [261] 2.783826e+18 2.788860e+18 2.793894e+18 2.798928e+18 2.803962e+18
#> [266] 2.808996e+18 2.814030e+18 2.819064e+18 2.824098e+18 2.829132e+18
#> [271] 2.834166e+18 2.839200e+18 2.844234e+18 2.849268e+18 2.854303e+18
#> [276] 2.859337e+18 2.864371e+18 2.869405e+18 2.874439e+18 2.879473e+18
#> [281] 2.884507e+18 2.889541e+18 2.894575e+18 2.899609e+18 2.904643e+18
#> [286] 2.909677e+18 2.914711e+18 2.919745e+18 2.924779e+18 2.929813e+18
#> [291] 2.934847e+18 2.939881e+18 2.944915e+18 2.949949e+18 2.954983e+18
#> [296] 2.960017e+18 2.965051e+18 2.970086e+18 2.975120e+18 2.980154e+18
#> [301] 2.985188e+18 2.990222e+18 2.995256e+18 3.000290e+18 3.005324e+18
#> [306] 3.010358e+18 3.015392e+18 3.020426e+18 3.025460e+18 3.030494e+18
#> [311] 3.035528e+18 3.040562e+18 3.045596e+18 3.050630e+18 3.055664e+18
#> [316] 3.060698e+18 3.065732e+18 3.070766e+18 3.075800e+18 3.080834e+18
#> [321] 3.085868e+18 3.090903e+18 3.095937e+18 3.100971e+18 3.106005e+18
#> [326] 3.111039e+18 3.116073e+18 3.121107e+18 3.126141e+18 3.131175e+18
#> [331] 3.136209e+18 3.141243e+18 3.146277e+18 3.151311e+18 3.156345e+18
#> [336] 3.161379e+18 3.166413e+18 3.171447e+18 3.176481e+18 3.181515e+18
#> [341] 3.186549e+18 3.191583e+18 3.196617e+18 3.201651e+18 3.206686e+18
#> [346] 3.211720e+18 3.216754e+18 3.221788e+18 3.226822e+18 3.231856e+18
#> [351] 3.236890e+18 3.241924e+18 3.246958e+18 3.251992e+18 3.257026e+18
#> [356] 3.262060e+18 3.267094e+18 3.272128e+18 3.277162e+18 3.282196e+18
#> [361] 3.287230e+18 3.292264e+18 3.297298e+18 3.302332e+18 3.307366e+18
#> [366] 3.312400e+18 3.317434e+18 3.322469e+18 3.327503e+18 3.332537e+18
#> [371] 3.337571e+18 3.342605e+18 3.347639e+18 3.352673e+18 3.357707e+18
#> [376] 3.362741e+18 3.367775e+18 3.372809e+18 3.377843e+18 3.382877e+18
#> [381] 3.387911e+18 3.392945e+18 3.397979e+18 3.403013e+18 3.408047e+18
#> [386] 3.413081e+18 3.418115e+18 3.423149e+18 3.428183e+18 3.433217e+18
#> [391] 3.438252e+18 3.443286e+18 3.448320e+18 3.453354e+18 3.458388e+18
#> [396] 3.463422e+18 3.468456e+18 3.473490e+18 3.478524e+18 3.483558e+18
#> [401] 3.488592e+18 3.493626e+18 3.498660e+18 3.503694e+18 3.508728e+18
#> [406] 3.513762e+18 3.518796e+18 3.523830e+18 3.528864e+18 3.533898e+18
#> [411] 3.538932e+18 3.543966e+18 3.549000e+18 3.554035e+18 3.559069e+18
#> [416] 3.564103e+18 3.569137e+18 3.574171e+18 3.579205e+18 3.584239e+18
#> [421] 3.589273e+18 3.594307e+18 3.599341e+18 3.604375e+18 3.609409e+18
#> [426] 3.614443e+18 3.619477e+18 3.624511e+18 3.629545e+18 3.634579e+18
#> [431] 3.639613e+18 3.644647e+18 3.649681e+18 3.654715e+18 3.659749e+18
#> [436] 3.664783e+18 3.669818e+18 3.674852e+18 3.679886e+18 3.684920e+18
#> [441] 3.689954e+18 3.694988e+18 3.700022e+18 3.705056e+18 3.710090e+18
#> [446] 3.715124e+18 3.720158e+18 3.725192e+18 3.730226e+18 3.735260e+18
#> [451] 3.740294e+18 3.745328e+18 3.750362e+18 3.755396e+18 3.760430e+18
#> [456] 3.765464e+18 3.770498e+18 3.775532e+18 3.780566e+18 3.785601e+18
#> [461] 3.790635e+18 3.795669e+18 3.800703e+18 3.805737e+18 3.810771e+18
#> [466] 3.815805e+18 3.820839e+18 3.825873e+18 3.830907e+18 3.835941e+18
#> [471] 3.840975e+18 3.846009e+18 3.851043e+18 3.856077e+18 3.861111e+18
#> [476] 3.866145e+18 3.871179e+18 3.876213e+18 3.881247e+18 3.886281e+18
#> [481] 3.891315e+18 3.896349e+18 3.901383e+18 3.906418e+18 3.911452e+18
#> [486] 3.916486e+18 3.921520e+18 3.926554e+18 3.931588e+18 3.936622e+18
#> [491] 3.941656e+18 3.946690e+18 3.951724e+18 3.956758e+18 3.961792e+18
#> [496] 3.966826e+18 3.971860e+18 3.976894e+18 3.981928e+18 3.986962e+18
#> [501] 3.991996e+18 3.997030e+18 4.002064e+18 4.007098e+18 4.012132e+18
#> [506] 4.017166e+18 4.022201e+18 4.027235e+18

with(sun.data, e2quantum_multipliers(w.length, molar = TRUE))

#>   [1] 2.449253e-06 2.457612e-06 2.465971e-06 2.474331e-06 2.482690e-06
#>   [6] 2.491049e-06 2.499408e-06 2.507767e-06 2.516127e-06 2.524486e-06
#>  [11] 2.532845e-06 2.541204e-06 2.549564e-06 2.557923e-06 2.566282e-06
#>  [16] 2.574641e-06 2.583000e-06 2.591360e-06 2.599719e-06 2.608078e-06
#>  [21] 2.616437e-06 2.624797e-06 2.633156e-06 2.641515e-06 2.649874e-06
#>  [26] 2.658233e-06 2.666593e-06 2.674952e-06 2.683311e-06 2.691670e-06
#>  [31] 2.700030e-06 2.708389e-06 2.716748e-06 2.725107e-06 2.733466e-06
#>  [36] 2.741826e-06 2.750185e-06 2.758544e-06 2.766903e-06 2.775263e-06
#>  [41] 2.783622e-06 2.791981e-06 2.800340e-06 2.808699e-06 2.817059e-06
#>  [46] 2.825418e-06 2.833777e-06 2.842136e-06 2.850496e-06 2.858855e-06
#>  [51] 2.867214e-06 2.875573e-06 2.883933e-06 2.892292e-06 2.900651e-06
#>  [56] 2.909010e-06 2.917369e-06 2.925729e-06 2.934088e-06 2.942447e-06
#>  [61] 2.950806e-06 2.959166e-06 2.967525e-06 2.975884e-06 2.984243e-06
#>  [66] 2.992602e-06 3.000962e-06 3.009321e-06 3.017680e-06 3.026039e-06
#>  [71] 3.034399e-06 3.042758e-06 3.051117e-06 3.059476e-06 3.067835e-06
#>  [76] 3.076195e-06 3.084554e-06 3.092913e-06 3.101272e-06 3.109632e-06
#>  [81] 3.117991e-06 3.126350e-06 3.134709e-06 3.143068e-06 3.151428e-06
#>  [86] 3.159787e-06 3.168146e-06 3.176505e-06 3.184865e-06 3.193224e-06
#>  [91] 3.201583e-06 3.209942e-06 3.218302e-06 3.226661e-06 3.235020e-06
#>  [96] 3.243379e-06 3.251738e-06 3.260098e-06 3.268457e-06 3.276816e-06
#> [101] 3.285175e-06 3.293535e-06 3.301894e-06 3.310253e-06 3.318612e-06
#> [106] 3.326971e-06 3.335331e-06 3.343690e-06 3.352049e-06 3.360408e-06
#> [111] 3.368768e-06 3.377127e-06 3.385486e-06 3.393845e-06 3.402204e-06
#> [116] 3.410564e-06 3.418923e-06 3.427282e-06 3.435641e-06 3.444001e-06
#> [121] 3.452360e-06 3.460719e-06 3.469078e-06 3.477437e-06 3.485797e-06
#> [126] 3.494156e-06 3.502515e-06 3.510874e-06 3.519234e-06 3.527593e-06
#> [131] 3.535952e-06 3.544311e-06 3.552670e-06 3.561030e-06 3.569389e-06
#> [136] 3.577748e-06 3.586107e-06 3.594467e-06 3.602826e-06 3.611185e-06
#> [141] 3.619544e-06 3.627904e-06 3.636263e-06 3.644622e-06 3.652981e-06
#> [146] 3.661340e-06 3.669700e-06 3.678059e-06 3.686418e-06 3.694777e-06
#> [151] 3.703137e-06 3.711496e-06 3.719855e-06 3.728214e-06 3.736573e-06
#> [156] 3.744933e-06 3.753292e-06 3.761651e-06 3.770010e-06 3.778370e-06
#> [161] 3.786729e-06 3.795088e-06 3.803447e-06 3.811806e-06 3.820166e-06
#> [166] 3.828525e-06 3.836884e-06 3.845243e-06 3.853603e-06 3.861962e-06
#> [171] 3.870321e-06 3.878680e-06 3.887039e-06 3.895399e-06 3.903758e-06
#> [176] 3.912117e-06 3.920476e-06 3.928836e-06 3.937195e-06 3.945554e-06
#> [181] 3.953913e-06 3.962273e-06 3.970632e-06 3.978991e-06 3.987350e-06
#> [186] 3.995709e-06 4.004069e-06 4.012428e-06 4.020787e-06 4.029146e-06
#> [191] 4.037506e-06 4.045865e-06 4.054224e-06 4.062583e-06 4.070942e-06
#> [196] 4.079302e-06 4.087661e-06 4.096020e-06 4.104379e-06 4.112739e-06
#> [201] 4.121098e-06 4.129457e-06 4.137816e-06 4.146175e-06 4.154535e-06
#> [206] 4.162894e-06 4.171253e-06 4.179612e-06 4.187972e-06 4.196331e-06
#> [211] 4.204690e-06 4.213049e-06 4.221408e-06 4.229768e-06 4.238127e-06
#> [216] 4.246486e-06 4.254845e-06 4.263205e-06 4.271564e-06 4.279923e-06
#> [221] 4.288282e-06 4.296641e-06 4.305001e-06 4.313360e-06 4.321719e-06
#> [226] 4.330078e-06 4.338438e-06 4.346797e-06 4.355156e-06 4.363515e-06
#> [231] 4.371875e-06 4.380234e-06 4.388593e-06 4.396952e-06 4.405311e-06
#> [236] 4.413671e-06 4.422030e-06 4.430389e-06 4.438748e-06 4.447108e-06
#> [241] 4.455467e-06 4.463826e-06 4.472185e-06 4.480544e-06 4.488904e-06
#> [246] 4.497263e-06 4.505622e-06 4.513981e-06 4.522341e-06 4.530700e-06
#> [251] 4.539059e-06 4.547418e-06 4.555777e-06 4.564137e-06 4.572496e-06
#> [256] 4.580855e-06 4.589214e-06 4.597574e-06 4.605933e-06 4.614292e-06
#> [261] 4.622651e-06 4.631010e-06 4.639370e-06 4.647729e-06 4.656088e-06
#> [266] 4.664447e-06 4.672807e-06 4.681166e-06 4.689525e-06 4.697884e-06
#> [271] 4.706244e-06 4.714603e-06 4.722962e-06 4.731321e-06 4.739680e-06
#> [276] 4.748040e-06 4.756399e-06 4.764758e-06 4.773117e-06 4.781477e-06
#> [281] 4.789836e-06 4.798195e-06 4.806554e-06 4.814913e-06 4.823273e-06
#> [286] 4.831632e-06 4.839991e-06 4.848350e-06 4.856710e-06 4.865069e-06
#> [291] 4.873428e-06 4.881787e-06 4.890146e-06 4.898506e-06 4.906865e-06
#> [296] 4.915224e-06 4.923583e-06 4.931943e-06 4.940302e-06 4.948661e-06
#> [301] 4.957020e-06 4.965379e-06 4.973739e-06 4.982098e-06 4.990457e-06
#> [306] 4.998816e-06 5.007176e-06 5.015535e-06 5.023894e-06 5.032253e-06
#> [311] 5.040612e-06 5.048972e-06 5.057331e-06 5.065690e-06 5.074049e-06
#> [316] 5.082409e-06 5.090768e-06 5.099127e-06 5.107486e-06 5.115846e-06
#> [321] 5.124205e-06 5.132564e-06 5.140923e-06 5.149282e-06 5.157642e-06
#> [326] 5.166001e-06 5.174360e-06 5.182719e-06 5.191079e-06 5.199438e-06
#> [331] 5.207797e-06 5.216156e-06 5.224515e-06 5.232875e-06 5.241234e-06
#> [336] 5.249593e-06 5.257952e-06 5.266312e-06 5.274671e-06 5.283030e-06
#> [341] 5.291389e-06 5.299748e-06 5.308108e-06 5.316467e-06 5.324826e-06
#> [346] 5.333185e-06 5.341545e-06 5.349904e-06 5.358263e-06 5.366622e-06
#> [351] 5.374981e-06 5.383341e-06 5.391700e-06 5.400059e-06 5.408418e-06
#> [356] 5.416778e-06 5.425137e-06 5.433496e-06 5.441855e-06 5.450215e-06
#> [361] 5.458574e-06 5.466933e-06 5.475292e-06 5.483651e-06 5.492011e-06
#> [366] 5.500370e-06 5.508729e-06 5.517088e-06 5.525448e-06 5.533807e-06
#> [371] 5.542166e-06 5.550525e-06 5.558884e-06 5.567244e-06 5.575603e-06
#> [376] 5.583962e-06 5.592321e-06 5.600681e-06 5.609040e-06 5.617399e-06
#> [381] 5.625758e-06 5.634117e-06 5.642477e-06 5.650836e-06 5.659195e-06
#> [386] 5.667554e-06 5.675914e-06 5.684273e-06 5.692632e-06 5.700991e-06
#> [391] 5.709350e-06 5.717710e-06 5.726069e-06 5.734428e-06 5.742787e-06
#> [396] 5.751147e-06 5.759506e-06 5.767865e-06 5.776224e-06 5.784583e-06
#> [401] 5.792943e-06 5.801302e-06 5.809661e-06 5.818020e-06 5.826380e-06
#> [406] 5.834739e-06 5.843098e-06 5.851457e-06 5.859817e-06 5.868176e-06
#> [411] 5.876535e-06 5.884894e-06 5.893253e-06 5.901613e-06 5.909972e-06
#> [416] 5.918331e-06 5.926690e-06 5.935050e-06 5.943409e-06 5.951768e-06
#> [421] 5.960127e-06 5.968486e-06 5.976846e-06 5.985205e-06 5.993564e-06
#> [426] 6.001923e-06 6.010283e-06 6.018642e-06 6.027001e-06 6.035360e-06
#> [431] 6.043719e-06 6.052079e-06 6.060438e-06 6.068797e-06 6.077156e-06
#> [436] 6.085516e-06 6.093875e-06 6.102234e-06 6.110593e-06 6.118952e-06
#> [441] 6.127312e-06 6.135671e-06 6.144030e-06 6.152389e-06 6.160749e-06
#> [446] 6.169108e-06 6.177467e-06 6.185826e-06 6.194186e-06 6.202545e-06
#> [451] 6.210904e-06 6.219263e-06 6.227622e-06 6.235982e-06 6.244341e-06
#> [456] 6.252700e-06 6.261059e-06 6.269419e-06 6.277778e-06 6.286137e-06
#> [461] 6.294496e-06 6.302855e-06 6.311215e-06 6.319574e-06 6.327933e-06
#> [466] 6.336292e-06 6.344652e-06 6.353011e-06 6.361370e-06 6.369729e-06
#> [471] 6.378088e-06 6.386448e-06 6.394807e-06 6.403166e-06 6.411525e-06
#> [476] 6.419885e-06 6.428244e-06 6.436603e-06 6.444962e-06 6.453321e-06
#> [481] 6.461681e-06 6.470040e-06 6.478399e-06 6.486758e-06 6.495118e-06
#> [486] 6.503477e-06 6.511836e-06 6.520195e-06 6.528554e-06 6.536914e-06
#> [491] 6.545273e-06 6.553632e-06 6.561991e-06 6.570351e-06 6.578710e-06
#> [496] 6.587069e-06 6.595428e-06 6.603788e-06 6.612147e-06 6.620506e-06
#> [501] 6.628865e-06 6.637224e-06 6.645584e-06 6.653943e-06 6.662302e-06
#> [506] 6.670661e-06 6.679021e-06 6.687380e-06