Energy fluence

This function returns the energy fluence for a given waveband of a light source spectrum given the duration of the exposure.

e_fluence(spct, w.band, exposure.time, wb.trim, use.cached.mult, use.hinges,
  allow.scaled, ...)

# S3 method for default
e_fluence(spct, w.band, exposure.time, wb.trim,
  use.cached.mult, use.hinges, allow.scaled, ...)

# S3 method for source_spct
e_fluence(spct, w.band = NULL, exposure.time,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = NULL, allow.scaled = FALSE, ...)

# S3 method for source_mspct
e_fluence(spct, w.band = NULL, exposure.time,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = NULL, allow.scaled = FALSE, ...,
  idx = !is.null(names(spct)))

Arguments

spct
an R object

w.band
a list of waveband objects or a waveband object

exposure.time
lubridate::duration

wb.trim
logical if TRUE wavebands crossing spectral data boundaries are trimmed, if FALSE, they are discarded

use.cached.mult
logical indicating whether multiplier values should be cached between calls

use.hinges
logical indicating whether to use hinges to reduce interpolation errors

allow.scaled
logical indicating whether scaled or normalized spectra as argument to spct are flagged as an error

...
other arguments (possibly ignored)

idx
logical whether to add a column with the names of the elements of spct

Value

One numeric value for each waveband with no change in scale factor, with name attribute set to the name of each waveband unless a named list is supplied in which case the names of the list elements are used. The exposure.time is copied to the output as an attribute. Units are as follows: (J) joules per exposure.

Methods (by class)

default: Default for generic function

source_spct: Calculate energy fluence from a source_spct object and the duration of the exposure.

source_mspct: Calculates energy fluence from a source_mspct object.

Note

The last two parameters control speed optimizations. The defaults should be suitable in mosts cases. If you will use repeatedly the same SWFs on many spectra measured at exactly the same wavelengths you may obtain some speed up by setting use.cached.mult=TRUE. However, be aware that you are responsible for ensuring that the wavelengths are the same in each call, as the only test done is for the length of the w.length vector.

See also

Other irradiance functions: e_irrad, energy_irradiance, fluence, irradiance, irrad, photon_irradiance, q_fluence, q_irrad

Examples

library(lubridate)
e_fluence(sun.spct, w.band = waveband(c(400,700)),
          exposure.time = lubridate::duration(3, "minutes") )

#>  range.400.700 
#>       35394.17 
#> attr(,"radiation.unit")
#> [1] "energy fluence (J m-2)"
#> attr(,"exposure.duration")
#> [1] "180s (~3 minutes)"