Energy irradiance

This function returns the energy irradiance for a given waveband of a light source spectrum.

e_irrad(spct, w.band, quantity, time.unit, wb.trim, use.cached.mult, use.hinges,
  allow.scaled, ...)

# S3 method for default
e_irrad(spct, w.band, quantity, time.unit, wb.trim,
  use.cached.mult, use.hinges, allow.scaled, ...)

# S3 method for source_spct
e_irrad(spct, w.band = NULL, quantity = "total",
  time.unit = NULL, wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.cached.mult = getOption("photobiology.use.cached.mult",
  default = FALSE), use.hinges = NULL, allow.scaled = !quantity %in%
  c("average", "mean", "total"), ...)

# S3 method for source_mspct
e_irrad(spct, w.band = NULL, quantity = "total",
  time.unit = NULL, wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.cached.mult = getOption("photobiology.use.cached.mult",
  default = FALSE), use.hinges = NULL, allow.scaled = !quantity %in%
  c("average", "mean", "total"), ..., idx = !is.null(names(spct)))

Arguments

spct
an R object

w.band
a list of waveband objects or a waveband object

quantity
character string One of "total", "average" or "mean", "contribution", "contribution.pc", "relative" or "relative.pc"

time.unit
character or lubridate::duration

wb.trim
logical if TRUE wavebands crossing spectral data boundaries are trimmed, if FALSE, they are discarded

use.cached.mult
logical indicating whether multiplier values should be cached between calls

use.hinges
logical indicating whether to use hinges to reduce interpolation errors

allow.scaled
logical indicating whether scaled or normalized spectra as argument to spct are flagged as an error

...
other arguments (possibly ignored)

idx
logical whether to add a column with the names of the elements of spct

Value

One numeric value for each waveband with no change in scale factor, with name attribute set to the name of each waveband unless a named list is supplied in which case the names of the list elements are used. The time.unit attribute is copied from the spectrum object to the output. Units are as follows: If units are absolute and time.unit is second, [W m-2 nm-1] -> [W m-2] If time.unit is day, [J d-1 m-2 nm-1] -> [J m-2]; if units are relative, fraction of one or percent.

Methods (by class)

default: Default for generic function

source_spct: Calculates energy irradiance from a source_spct object.

source_mspct: Calculates energy irradiance from a source_mspct object.

Note

The last two parameters control speed optimizations. The defaults should be suitable in mosts cases. If you will use repeatedly the same SWFs on many spectra measured at exactly the same wavelengths you may obtain some speed up by setting use.cached.mult=TRUE. However, be aware that you are responsible for ensuring that the wavelengths are the same in each call, as the only test done is for the length of the w.length vector.

See also

Other irradiance functions: e_fluence, energy_irradiance, fluence, irradiance, irrad, photon_irradiance, q_fluence, q_irrad

Examples

e_irrad(sun.spct, waveband(c(400,700)))

#>  range.400.700 
#>       196.6343 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance total"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3))

#>  range.400.500  range.500.600  range.600.700 
#>       69.69043       68.48950       58.45434 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance total"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "total")

#>  range.400.500  range.500.600  range.600.700 
#>       69.69043       68.48950       58.45434 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance total"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "average")

#>  range.400.500  range.500.600  range.600.700 
#>      0.6969043      0.6848950      0.5845434 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance average"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "relative")

#>  range.400.500  range.500.600  range.600.700 
#>      0.3544165      0.3483091      0.2972744 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance relative"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "relative.pc")

#>  range.400.500  range.500.600  range.600.700 
#>       35.44165       34.83091       29.72744 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance relative.pc"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "contribution")

#>  range.400.500  range.500.600  range.600.700 
#>      0.2589520      0.2544897      0.2172015 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance contribution"

e_irrad(sun.spct, split_bands(c(400,700), length.out = 3),
        quantity = "contribution.pc")

#>  range.400.500  range.500.600  range.600.700 
#>       25.89520       25.44897       21.72015 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy irradiance contribution.pc"