Energy-based photo-response

This function returns the mean, total, or contribution of response for each waveband and a response spectrum.

e_response(spct, w.band, quantity, time.unit, wb.trim, use.hinges, ...)

# S3 method for default
e_response(spct, w.band, quantity, time.unit, wb.trim,
  use.hinges, ...)

# S3 method for response_spct
e_response(spct, w.band = NULL, quantity = "total",
  time.unit = NULL, wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.hinges = getOption("photobiology.use.hinges", default =
  NULL), ...)

# S3 method for response_mspct
e_response(spct, w.band = NULL, quantity = "total",
  time.unit = NULL, wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.hinges = getOption("photobiology.use.hinges", default =
  NULL), ..., idx = !is.null(names(spct)))

Arguments

spct
an R object

w.band
waveband or list of waveband objects or a numeric vector of length two. The waveband(s) determine the region(s) of the spectrum that are summarized. If a numeric range is supplied a waveband object is constructed on the fly from it.

quantity
character string One of "total", "average" or "mean", "contribution", "contribution.pc", "relative" or "relative.pc"

time.unit
character or lubridate::duration

wb.trim
logical if TRUE wavebands crossing spectral data boundaries are trimmed, if FALSE, they are discarded

use.hinges
logical indicating whether to use hinges to reduce interpolation errors

...
other arguments

idx
logical whether to add a column with the names of the elements of spct

Value

A single numeric value expressed either as a fraction of one or a percentage, or a vector of the same length as the list of wave.bands. The quantity returned, although always on energy-based units, depends on the value of quantity.

Methods (by class)

default: Default method for generic function

response_spct: Method for response spectra.

response_mspct: Calculates energy response from a response_mspct

Note

The parameter use.hinges controls speed optimization. The defaults should be suitable in mosts cases. Only the range of wavelengths in the wavebands is used and all BSWFs are ignored.

See also

Other response functions: q_response, response

Examples

e_response(ccd.spct, new_waveband(200,300))

#>  range.200.300.tr.lo 
#>          0.000115335 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy response total"

e_response(photodiode.spct)

#>    Total 
#> 24.40478 
#> attr(,"time.unit")
#> [1] "second"
#> attr(,"radiation.unit")
#> [1] "energy response total"