Extract or replace parts of a spectrum

Just like extraction and replacement with indexes in base R, but preserving the special attributes used in spectral classes and checking for validity of remaining spectral data.

# S3 method for generic_spct
[(x, i, j, drop = NULL)

# S3 method for raw_spct
[(x, i, j, drop = NULL)

# S3 method for cps_spct
[(x, i, j, drop = NULL)

# S3 method for source_spct
[(x, i, j, drop = NULL)

# S3 method for response_spct
[(x, i, j, drop = NULL)

# S3 method for filter_spct
[(x, i, j, drop = NULL)

# S3 method for reflector_spct
[(x, i, j, drop = NULL)

# S3 method for object_spct
[(x, i, j, drop = NULL)

# S3 method for chroma_spct
[(x, i, j, drop = NULL)

# S3 method for generic_spct
[(x, i, j) <- value

# S3 method for generic_spct
$(x, name) <- value

Arguments

x
spectral object from which to extract element(s) or in which to replace element(s)

i
index for rows,

j
index for columns, specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Please, see Extract for more details.

drop
logical. If TRUE the result is coerced to the lowest possible dimension. The default is FALSE unless the result is a single column.

value
A suitable replacement value: it will be repeated a whole number of times if necessary and it may be coerced: see the Coercion section. If NULL, deletes the column if a single column is selected.

name
A literal character string or a name (possibly backtick quoted). For extraction, this is normally (see under ‘Environments’) partially matched to the names of the object.

Value

An object of the same class as x but containing only the subset of rows and columns that are selected. See details for special cases.

Details

These methods are just wrappers on the method for data.frame objects which copy the additional attributes used by these classes, and validate the extracted object as a spectral object. When drop is TRUE and the returned object has only one column, then a vector is returned. If the extrated columns are more than one but do not include w.length, a data frame is returned instead of a spectral object.

See also

subset.data.frame and trim_spct

Examples

sun.spct[sun.spct$w.length > 400, ]

#> Object: source_spct [400 x 3]
#> Wavelength range 401 to 800 nm, step 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 400 × 3
#>    w.length s.e.irrad    s.q.irrad
#> *     <dbl>     <dbl>        <dbl>
#> 1       401 0.6261742 2.098967e-06
#> 2       402 0.6497388 2.183388e-06
#> 3       403 0.6207287 2.091091e-06
#> 4       404 0.6370489 2.151395e-06
#> 5       405 0.6263786 2.120596e-06
#> 6       406 0.5996440 2.035099e-06
#> 7       407 0.5951226 2.024729e-06
#> 8       408 0.6256209 2.133720e-06
#> 9       409 0.6615421 2.261761e-06
#> 10      410 0.5605505 1.921165e-06
#> # ... with 390 more rows

subset(sun.spct, w.length > 400)

#> Object: source_spct [400 x 3]
#> Wavelength range 401 to 800 nm, step 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 400 × 3
#>    w.length s.e.irrad    s.q.irrad
#> *     <dbl>     <dbl>        <dbl>
#> 1       401 0.6261742 2.098967e-06
#> 2       402 0.6497388 2.183388e-06
#> 3       403 0.6207287 2.091091e-06
#> 4       404 0.6370489 2.151395e-06
#> 5       405 0.6263786 2.120596e-06
#> 6       406 0.5996440 2.035099e-06
#> 7       407 0.5951226 2.024729e-06
#> 8       408 0.6256209 2.133720e-06
#> 9       409 0.6615421 2.261761e-06
#> 10      410 0.5605505 1.921165e-06
#> # ... with 390 more rows

tmp.spct <- sun.spct
tmp.spct[tmp.spct$s.e.irrad < 1e-5 , "s.e.irrad"] <- 0
e2q(tmp.spct[ , c("w.length", "s.e.irrad")]) # restore data consistency!

#> Object: source_spct [522 x 3]
#> Wavelength range 280 to 800 nm, step 0.9230769 to 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 522 × 3
#>    w.length s.e.irrad s.q.irrad
#>       <dbl>     <dbl>     <dbl>
#> 1  280.0000         0         0
#> 2  280.9231         0         0
#> 3  281.8462         0         0
#> 4  282.7692         0         0
#> 5  283.6923         0         0
#> 6  284.6154         0         0
#> 7  285.5385         0         0
#> 8  286.4615         0         0
#> 9  287.3846         0         0
#> 10 288.3077         0         0
#> # ... with 512 more rows