Counts per second from a measurement of a plastic film

A dataset containing the wavelengths at a varying interval and counts per second data. Two spectra are included a reference measurement of the light source plus a measurement of a plastic film.

Format

A filter_cps.mspct object with 611 rows and 2 variables

Details

A collection of two spectra as a cps_mspct object, each each spectrum being a cps_spct object with two columns:

w.length (nm).

cps (counts/s)

See also

Other Spectral data examples: A.illuminant.spct, D65.illuminant.spct, black_body.spct, ccd.spct, clear.spct, clear_body.spct, green_leaf.spct, opaque.spct, photodiode.spct, polyester.spct, sun.daily.data, sun.daily.spct, sun.data, sun.spct, white_body.spct, white_led.cps_spct, white_led.raw_spct, white_led.source_spct, yellow_gel.spct