Collection-of-spectra constructor

Converts a list of spectral objects into a "multi spectrum" object by setting the class attribute of the list of spectra to the corresponding multi-spct class, check that components of the list belong to the expected class.

generic_mspct(l = NULL, class = "generic_spct", ncol = 1, byrow = FALSE,
  dim = c(length(l)%/%ncol, ncol))

calibration_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

raw_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

cps_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

source_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

filter_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

reflector_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

object_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

response_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

chroma_mspct(l = NULL, ncol = 1, byrow = FALSE, ...)

Arguments

l
list of generic_spct or derived classes

class
character The multi spectrum object class or the expected class for the elements of l

ncol
integer Number of 'virtual' columns in data

byrow
logical If ncol > 1 how to read in the data

dim
integer vector of dimensions

...
ignored

Functions

calibration_mspct: Specialization for collections of calibration_spct objects.

raw_mspct: Specialization for collections of raw_spct objects.

cps_mspct: Specialization for collections of cps_spct objects.

source_mspct: Specialization for collections of source_spct objects.

filter_mspct: Specialization for collections of filter_spct objects.

reflector_mspct: Specialization for collections of reflector_spct objects.

object_mspct: Specialization for collections of object_spct objects.

response_mspct: Specialization for collections of response_spct objects.

chroma_mspct: Specialization for collections of chroma_spct objects.

Note

Setting class = source_spct or class = source_mspct makes no difference

Examples

filter_mspct(list(polyester.spct, yellow_gel.spct))

#> Object: filter_mspct [2 x 1]
#> --- Member: spct_1 ---
#> Object: filter_spct [611 x 2]
#> Wavelength range 190 to 800 nm, step 1 nm 
#> Label: clear polyester film, 125um thick 
#> 
#> # A tibble: 611 x 2
#>    w.length    Tfr
#>       <int>  <dbl>
#>  1      190 0.0110
#>  2      191 0.0100
#>  3      192 0.0110
#>  4      193 0.0150
#>  5      194 0.0160
#>  6      195 0.0120
#>  7      196 0.0110
#>  8      197 0.0110
#>  9      198 0.0110
#> 10      199 0.0110
#> # ... with 601 more rows
#> --- Member: spct_2 ---
#> Object: filter_spct [611 x 2]
#> Wavelength range 190 to 800 nm, step 1 nm 
#> Label: yellow theatrical 'gel', Rosco supergel no. 312, 'canary yellow' 
#> 
#> # A tibble: 611 x 2
#>    w.length       Tfr
#>       <int>     <dbl>
#>  1      190 0.00100  
#>  2      191 0.00350  
#>  3      192 0.00400  
#>  4      193 0.0000100
#>  5      194 0.0000100
#>  6      195 0.0000100
#>  7      196 0.0000100
#>  8      197 0.0000100
#>  9      198 0.0000100
#> 10      199 0.0000100
#> # ... with 601 more rows
#> 
#> --- END ---