Get peaks and valleys in a spectrum

These functions find peaks (local maxima) or valleys (local minima) in a spectrum, using a user selectable size threshold relative to the tallest peak (global maximum). This a wrapper built on top of function peaks from package splus2R.

get_peaks(x, y, ignore_threshold = 0, span = 5, strict = TRUE,
  x_unit = "", x_digits = 3)

get_valleys(x, y, ignore_threshold = 0, span = 5, strict = TRUE,
  x_unit = "", x_digits = 3)

Arguments

x
numeric

y
numeric

ignore_threshold
numeric Value between 0.0 and 1.0 indicating the relative size compared to tallest peak or deepest valley of the peaks to return.

span
numeric A peak is defined as an element in a sequence which is greater than all other elements within a window of width span centered at that element. For example, a value of 3 means that a peak is bigger than both of its neighbors.

strict
logical Flag: if TRUE, an element must be strictly greater than all other values in its window to be considered a peak. Default: TRUE.

x_unit
character Vector of texts to be pasted at end of labels built from x value at peaks.

x_digits
numeric Number of significant digits in wavelength label.

Value

A data frame with variables w.length and s.irrad with their values at the peaks or valleys plus a character variable of labels.

See also

Other peaks and valleys functions: find_peaks, peaks, valleys

Examples

with(sun.spct, get_peaks(w.length, s.e.irrad))

#>      x         y label
#> 1  321 0.1822031   321
#> 2  330 0.3295190   330
#> 3  335 0.3129253   335
#> 4  340 0.3352353   340
#> 5  343 0.3380052   343
#> 6  347 0.3207918   347
#> 7  350 0.3453572   350
#> 8  354 0.3758625   354
#> 9  360 0.3707068   360
#> 10 366 0.4491898   366
#> 11 370 0.4393233   370
#> 12 378 0.4969714   378
#> 13 381 0.4362110   381
#> 14 385 0.3915446   385
#> 15 391 0.4822105   391
#> 16 395 0.4699886   395
#> 17 402 0.6497388   402
#> 18 409 0.6615421   409
#> 19 412 0.6742498   412
#> 20 416 0.6761818   416
#> 21 421 0.6701269   421
#> 22 426 0.6388873   426
#> 23 436 0.7336607   436
#> 24 442 0.7188581   442
#> 25 451 0.8204633   451
#> 26 457 0.7984935   457
#> 27 461 0.7711277   461
#> 28 468 0.7665312   468
#> 29 472 0.7693357   472
#> 30 475 0.7724634   475
#> 31 478 0.7869773   478
#> 32 482 0.7832759   482
#> 33 490 0.7728111   490
#>  [ reached getOption("max.print") -- omitted 43 rows ]

with(sun.spct, get_valleys(w.length, s.e.irrad))

#>      x         y label
#> 1  316 0.1020587   316
#> 2  323 0.1680533   323
#> 3  328 0.2648573   328
#> 4  331 0.2895510   331
#> 5  337 0.2609770   337
#> 6  341 0.3053287   341
#> 7  344 0.2675878   344
#> 8  349 0.2981687   349
#> 9  352 0.3234148   352
#> 10 358 0.2544907   358
#> 11 368 0.3980621   368
#> 12 374 0.3260057   374
#> 13 383 0.2614310   383
#> 14 393 0.2422023   393
#> 15 397 0.2541827   397
#> 16 403 0.6207287   403
#> 17 407 0.5951226   407
#> 18 410 0.5605505   410
#> 19 413 0.6456365   413
#> 20 420 0.5988416   420
#> 21 424 0.6026922   424
#> 22 431 0.4136900   431
#> 23 434 0.6298332   434
#> 24 439 0.6028502   439
#> 25 446 0.6652017   446
#> 26 453 0.6972674   453
#> 27 458 0.7618389   458
#> 28 467 0.7429891   467
#> 29 471 0.7424066   471
#> 30 474 0.7454852   474
#> 31 479 0.7536975   479
#> 32 487 0.6511654   487
#> 33 492 0.6896611   492
#>  [ reached getOption("max.print") -- omitted 42 rows ]