Return the First and Last Part of an Object

Returns the first and last parts of a spectrum, dataframe, vector, function, table or ftable.

head_tail(x, ...)

# S3 method for default
head_tail(x, n = 3L, ...)

# S3 method for data.frame
head_tail(x, n = 3L, ...)

# S3 method for matrix
head_tail(x, n = 3L, ...)

# S3 method for function
head_tail(x, n = 6L, ...)

# S3 method for table
head_tail(x, n = 6L, ...)

# S3 method for ftable
head_tail(x, n = 6L, ...)

Arguments

x
an object.

...
arguments to be passed to or from other methods.

n
a single integer. If positive, half the size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function, to be retained from each end. If negative, all but the n last and n first number of elements of x.

Value

An object (usually) like x but generally smaller. For ftable objects x, a transformed format(x).

Details

The value returned by head_tail() is equivalent to row binding the the values returned by head() and tail(), although not implemented in this way. The same specializations as defined in package 'utils' for head() and tail() have been implemented.

Methods (by class)

default:

data.frame:

matrix:

function:

table:

ftable:

Note

For some types of input, like functions, the output may be confusing, however, we have opted for consistency with existing functions. The code is in part a revision of that of head() and tail() from package utils. I have been missing this method especially when checking spectral data, as both ends are of interest. head_tail() methods for function, table and ftable classes, are wrappers for head() method.

See also

head, and compare the examples and the values returned to the examples below.

Examples

head_tail(letters)

#> [1] "a" "b" "c" "x" "y" "z"

head_tail(letters, n = -6L)

#>  [1] "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s" "t"

head_tail(freeny.x, n = 10L)

#>       lag quarterly revenue price index income level market potential
#>  [1,]               8.79636     4.70997      5.82110          12.9699
#>  [2,]               8.79236     4.70217      5.82558          12.9733
#>  [3,]               8.79137     4.68944      5.83112          12.9774
#>  [4,]               8.81486     4.68558      5.84046          12.9806
#>  [5,]               8.81301     4.64019      5.85036          12.9831
#>  [6,]               8.90751     4.62553      5.86464          12.9854
#>  [7,]               8.93673     4.61991      5.87769          12.9900
#>  [8,]               8.96161     4.61654      5.89763          12.9943
#>  [9,]               8.96044     4.61407      5.92574          12.9992
#> [10,]               9.00868     4.60766      5.94232          13.0033
#> [11,]               9.58123     4.38513      6.14705          13.1266
#> [12,]               9.60048     4.37320      6.15336          13.1356
#> [13,]               9.64496     4.32770      6.15627          13.1415
#> [14,]               9.64390     4.32023      6.16274          13.1444
#> [15,]               9.69405     4.30909      6.17369          13.1459
#> [16,]               9.69958     4.30909      6.16135          13.1520
#> [17,]               9.68683     4.30552      6.18231          13.1593
#> [18,]               9.71774     4.29627      6.18768          13.1579
#> [19,]               9.74924     4.27839      6.19377          13.1625
#> [20,]               9.77536     4.27789      6.20030          13.1664

head_tail(freeny.y)

#> [1] 8.79236 8.79137 8.81486 9.74924 9.77536 9.79424

head_tail(stats::ftable(Titanic))

#>                                                
#>                           "Survived" "No" "Yes"
#>  "Class" "Sex"    "Age"                        
#>  "1st"   "Male"   "Child"               0     5
#>                   "Adult"             118    57
#>          "Female" "Child"               0     1
#>                   "Adult"               4   140
#>  "2nd"   "Male"   "Child"               0    11
#>                   "Adult"             154    14
#>  "3rd"   "Male"   "Child"              35    13
#>                   "Adult"             387    75
#>          "Female" "Child"              17    14
#>                   "Adult"              89    76
#>  "Crew"  "Male"   "Child"               0     0
#>                   "Adult"             670   192
#>          "Female" "Child"               0     0
#>                   "Adult"               3    20