Insert wavelength values into spectral data.

Inserting wavelengths values immediately before and after a discontinuity in the SWF, greatly reduces the errors caused by interpolating the weighted irradiance during integration of the effective spectral irradiance. This is specially true when data have a large wavelength step size.

insert_hinges(x, y, h)

Arguments

x
numeric array (sorted in increasing order)

y
numeric array

h
a numeric array giving the wavelengths at which the y values should be inserted by interpolation, no interpolation is indicated by an empty array (numeric(0))

Value

a data.frame with variables x and y

Note

Insertion is a costly operation but I have tried to optimize this function as much as possible by avoiding loops. Earlier this function was implemented in C++, but a bug was discovered and I have now rewritten it using R.

Examples

with(sun.data,
    insert_hinges(w.length, s.e.irrad,
       c(399.99, 400.00, 699.99, 700.00)))

#>          x            y
#> 1   293.00 2.609665e-06
#> 2   294.00 6.142401e-06
#> 3   295.00 2.176175e-05
#> 4   296.00 6.780119e-05
#> 5   297.00 1.533491e-04
#> 6   298.00 3.669677e-04
#> 7   299.00 7.845430e-04
#> 8   300.00 1.264554e-03
#> 9   301.00 2.623718e-03
#> 10  302.00 3.922583e-03
#> 11  303.00 8.974134e-03
#> 12  304.00 1.165567e-02
#> 13  305.00 1.799127e-02
#> 14  306.00 2.076166e-02
#> 15  307.00 3.059158e-02
#> 16  308.00 3.997854e-02
#> 17  309.00 4.163908e-02
#> 18  310.00 4.871239e-02
#> 19  311.00 7.852756e-02
#> 20  312.00 8.036738e-02
#> 21  313.00 9.428946e-02
#> 22  314.00 1.054126e-01
#> 23  315.00 1.127901e-01
#> 24  316.00 1.020587e-01
#> 25  317.00 1.487690e-01
#> 26  318.00 1.413919e-01
#> 27  319.00 1.569692e-01
#> 28  320.00 1.741691e-01
#> 29  321.00 1.822031e-01
#> 30  322.00 1.787215e-01
#> 31  323.00 1.680533e-01
#> 32  324.00 2.075508e-01
#> 33  325.00 2.168055e-01
#> 34  326.00 2.774416e-01
#> 35  327.00 2.851096e-01
#> 36  328.00 2.648573e-01
#> 37  329.00 2.998745e-01
#> 38  330.00 3.295190e-01
#> 39  331.00 2.895510e-01
#> 40  332.00 3.000718e-01
#> 41  333.00 2.951501e-01
#> 42  334.00 2.936962e-01
#> 43  335.00 3.129253e-01
#> 44  336.00 2.712433e-01
#> 45  337.00 2.609770e-01
#> 46  338.00 2.904669e-01
#> 47  339.00 3.108404e-01
#> 48  340.00 3.352353e-01
#> 49  341.00 3.053287e-01
#> 50  342.00 3.221950e-01
#>  [ reached getOption("max.print") -- omitted 460 rows ]

with(sun.data,
    insert_hinges(w.length, s.e.irrad,
      c(100, 399.50, 399.99, 400.00, 699.99, 700.00, 799.99, 1000)))

#>          x            y
#> 1   293.00 2.609665e-06
#> 2   294.00 6.142401e-06
#> 3   295.00 2.176175e-05
#> 4   296.00 6.780119e-05
#> 5   297.00 1.533491e-04
#> 6   298.00 3.669677e-04
#> 7   299.00 7.845430e-04
#> 8   300.00 1.264554e-03
#> 9   301.00 2.623718e-03
#> 10  302.00 3.922583e-03
#> 11  303.00 8.974134e-03
#> 12  304.00 1.165567e-02
#> 13  305.00 1.799127e-02
#> 14  306.00 2.076166e-02
#> 15  307.00 3.059158e-02
#> 16  308.00 3.997854e-02
#> 17  309.00 4.163908e-02
#> 18  310.00 4.871239e-02
#> 19  311.00 7.852756e-02
#> 20  312.00 8.036738e-02
#> 21  313.00 9.428946e-02
#> 22  314.00 1.054126e-01
#> 23  315.00 1.127901e-01
#> 24  316.00 1.020587e-01
#> 25  317.00 1.487690e-01
#> 26  318.00 1.413919e-01
#> 27  319.00 1.569692e-01
#> 28  320.00 1.741691e-01
#> 29  321.00 1.822031e-01
#> 30  322.00 1.787215e-01
#> 31  323.00 1.680533e-01
#> 32  324.00 2.075508e-01
#> 33  325.00 2.168055e-01
#> 34  326.00 2.774416e-01
#> 35  327.00 2.851096e-01
#> 36  328.00 2.648573e-01
#> 37  329.00 2.998745e-01
#> 38  330.00 3.295190e-01
#> 39  331.00 2.895510e-01
#> 40  332.00 3.000718e-01
#> 41  333.00 2.951501e-01
#> 42  334.00 2.936962e-01
#> 43  335.00 3.129253e-01
#> 44  336.00 2.712433e-01
#> 45  337.00 2.609770e-01
#> 46  338.00 2.904669e-01
#> 47  339.00 3.108404e-01
#> 48  340.00 3.352353e-01
#> 49  341.00 3.053287e-01
#> 50  342.00 3.221950e-01
#>  [ reached getOption("max.print") -- omitted 462 rows ]