Gives irradiance from spectral irradiance.

This function gives the result of integrating spectral irradiance over wavelengths.

integrate_xy(x, y)

Arguments

x
numeric vector.

y
numeric vector.

Value

a single numeric value with no change in scale factor: e.g. [W m-2 nm-1] -> [W m-2]

See also

Other low-level functions operating on numeric vectors.: as_energy, as_quantum_mol, calc_multipliers, div_spectra, energy_irradiance, energy_ratio, insert_hinges, interpolate_spectrum, irradiance, l_insert_hinges, oper_spectra, photon_irradiance, photon_ratio, photons_energy_ratio, prod_spectra, s_e_irrad2rgb, split_energy_irradiance, split_photon_irradiance, subt_spectra, sum_spectra, trim_tails, v_insert_hinges

Examples

with(sun.data, integrate_xy(w.length, s.e.irrad))

#> [1] 269.1249