Map spectra to new wavelength values.

This function returns the result of interpolating spectral data from the original set of wavelengths to a new one.

interpolate_wl(x, w.length.out, fill, length.out, ...)

# S3 method for default
interpolate_wl(x, w.length.out, fill, length.out, ...)

# S3 method for generic_spct
interpolate_wl(x, w.length.out = NULL, fill = NA,
  length.out = NULL, ...)

# S3 method for generic_mspct
interpolate_wl(x, w.length.out = NULL, fill = NA,
  length.out = NULL, ...)

Arguments

x
an R object

w.length.out
numeric array of wavelengths (nm)

fill
a value to be assigned to out of range wavelengths

length.out
numeric value

...
not used

Value

A new spectral object of the same class as argument spct.

Details

If length.out it is a numeric value, then gives the number of rows in the output, if it is NULL, the values in the numeric vector w.length.out are used. If both are not NULL then the range of w.length.out and length.out are used to generate a vector of wavelength. A value of NULL for fill prevents extrapolation.

Methods (by class)

default: Default for generic function

generic_spct: Interpolate wavelength in an object of class "generic_spct" or derived.

generic_mspct: Interpolate wavelength in an object of class "generic_mspct" or derived.

Note

The default fill = NA fills extrpolated values with NA. Giving NULL as argument for fill deletes wavelengths outside the input data range from the returned spectrum. A numerical value can be also be provided as fill. This function calls interpolate_spectrum for each non-wavelength column in the input spectra object.

Examples

interpolate_wl(sun.spct, 400:500, NA)

#> Object: source_spct [101 x 3]
#> Wavelength range 400 to 500 nm, step 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 101 × 3
#>    w.length s.e.irrad    s.q.irrad
#>       <int>     <dbl>        <dbl>
#> 1       400 0.6081049 2.033314e-06
#> 2       401 0.6261742 2.098967e-06
#> 3       402 0.6497388 2.183388e-06
#> 4       403 0.6207287 2.091091e-06
#> 5       404 0.6370489 2.151395e-06
#> 6       405 0.6263786 2.120596e-06
#> 7       406 0.5996440 2.035099e-06
#> 8       407 0.5951226 2.024729e-06
#> 9       408 0.6256209 2.133720e-06
#> 10      409 0.6615421 2.261761e-06
#> # ... with 91 more rows

interpolate_wl(sun.spct, 400:500, NULL)

#> Object: source_spct [101 x 3]
#> Wavelength range 400 to 500 nm, step 1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 101 × 3
#>    w.length s.e.irrad    s.q.irrad
#>       <int>     <dbl>        <dbl>
#> 1       400 0.6081049 2.033314e-06
#> 2       401 0.6261742 2.098967e-06
#> 3       402 0.6497388 2.183388e-06
#> 4       403 0.6207287 2.091091e-06
#> 5       404 0.6370489 2.151395e-06
#> 6       405 0.6263786 2.120596e-06
#> 7       406 0.5996440 2.035099e-06
#> 8       407 0.5951226 2.024729e-06
#> 9       408 0.6256209 2.133720e-06
#> 10      409 0.6615421 2.261761e-06
#> # ... with 91 more rows

interpolate_wl(sun.spct, seq(200, 1000, by=0.1), 0)

#> Object: source_spct [8,001 x 3]
#> Wavelength range 200 to 1000 nm, step 0.1 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 8,001 × 3
#>    w.length s.e.irrad s.q.irrad
#>       <dbl>     <dbl>     <dbl>
#> 1     200.0         0         0
#> 2     200.1         0         0
#> 3     200.2         0         0
#> 4     200.3         0         0
#> 5     200.4         0         0
#> 6     200.5         0         0
#> 7     200.6         0         0
#> 8     200.7         0         0
#> 9     200.8         0         0
#> 10    200.9         0         0
#> # ... with 7,991 more rows

interpolate_wl(sun.spct, c(400,500), length.out=201)

#> Object: source_spct [201 x 3]
#> Wavelength range 400 to 500 nm, step 0.5 nm 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647354 E 
#> Time unit 1s
#> 
#> # A tibble: 201 × 3
#>    w.length s.e.irrad    s.q.irrad
#>       <dbl>     <dbl>        <dbl>
#> 1     400.0 0.6081049 2.033314e-06
#> 2     400.5 0.6171395 2.066140e-06
#> 3     401.0 0.6261742 2.098967e-06
#> 4     401.5 0.6379565 2.141177e-06
#> 5     402.0 0.6497388 2.183388e-06
#> 6     402.5 0.6352337 2.137239e-06
#> 7     403.0 0.6207287 2.091091e-06
#> 8     403.5 0.6288888 2.121243e-06
#> 9     404.0 0.6370489 2.151395e-06
#> 10    404.5 0.6317138 2.135996e-06
#> # ... with 191 more rows