Query if a spectrum contains photon- or energy-based data.

Functions to check if source_spct and response_spct objects contains photon-based or energy-based data.

is_photon_based(x)

is_energy_based(x)

Arguments

x
any R object

Value

is_photon_based returns TRUE if its argument is a a source_spct or a response_spct object that contains photon base data and FALSE if such an object does not contain such data, but returns NA for any other R object, including those belonging other generic_spct-derived classes. is_energy_based returns TRUE if its argument is a a source_spct or a response_spct object that contains energy base data and FALSE if such an object does not contain such data, but returns NA for any other R object, including those belonging other generic_spct-derived classes

See also

Other query units functions: is_absorbance_based

Examples

is_photon_based(sun.spct)

#> [1] TRUE

my.spct <- dplyr::select(sun.spct, w.length, s.e.irrad)
is.source_spct(my.spct)

#> [1] TRUE

is_photon_based(my.spct)

#> [1] FALSE

is_energy_based(sun.spct)

#> [1] TRUE

my.spct <- dplyr::select(sun.spct, w.length, s.q.irrad)
is.source_spct(my.spct)

#> [1] TRUE

is_energy_based(my.spct)

#> [1] FALSE