Join all spectra in a collection

Join all the spectra contained in a homogenous collection, returning a data frame with spectral-data columns named according to the names of the spectra in the collection. By default a full join is done, filling the spectral data for missing wave lengths in individual spectra with NA.

join_mspct(x, type, ...)

# S3 method for default
join_mspct(x, type = "full", ...)

# S3 method for generic_mspct
join_mspct(x, type = "full", col.name, ...)

# S3 method for source_mspct
join_mspct(x, type = "full", unit.out = "energy",
  ...)

# S3 method for response_mspct
join_mspct(x, type = "full", unit.out = "energy",
  ...)

# S3 method for filter_mspct
join_mspct(x, type = "full",
  qty.out = "transmittance", ...)

# S3 method for reflector_mspct
join_mspct(x, type = "full", ...)

# S3 method for object_mspct
join_mspct(x, type = "full", qty.out, ...)

Arguments

x
A generic_mspct object, or an object of a class derived from generic_mspct.

type
character Type of join: "left", "right", "inner" or "full" (default). See details for more information.

...
ignored (possibly used by derived methods).

col.name
character, name of the column in the spectra to be preserved, in addition to "w.length".

unit.out
character Allowed values "energy", and "photon", or its alias "quantum".

qty.out
character Allowed values "transmittance", and "absorbance".

Value

An object of class dataframe, with the spectra joined by wave length, with rows in addition sorted by wave length (variable w.length).

Methods (by class)

default:

generic_mspct:

source_mspct:

response_mspct:

filter_mspct:

reflector_mspct:

object_mspct:

Note

Currently only generic_spct, source_mspct, response_mspct, filter_mspct, reflector_mspct and object_mspct classes have this method implemented.