Insert wavelength values into spectral data.

Inserting wavelengths values immediately before and after a discontinuity in the SWF, greatly reduces the errors caused by interpolating the weighted irradiance during integration of the effective spectral irradiance. This is specially true when data have a large wavelength step size.

l_insert_hinges(x, y, h)

Arguments

x
numeric vector (sorted in increasing order)

y
numeric vector

h
a numeric vector giving the wavelengths at which the y values should be inserted by interpolation, no interpolation is indicated by an empty vector (numeric(0))

Value

a list with variables x and y. Unless the hinge values were already present in y, each inserted hinge, expands the vectors by two values.

See also