Photon:energy ratio

This function gives the photons:energy ratio between for one given waveband of a radiation spectrum.

photons_energy_ratio(w.length, s.irrad, w.band = NULL, unit.in = "energy",
  check.spectrum = TRUE, use.cached.mult = FALSE,
  use.hinges = getOption("photobiology.use.hinges", default = NULL))

Arguments

w.length
numeric vector of wavelength (nm).

s.irrad
numeric vector of spectral (energy) irradiances (W m-2 nm-1).

w.band
waveband object.

unit.in
character Allowed values "energy", and "photon", or its alias "quantum".

check.spectrum
logical Flag telling whether to sanity check input data, default is TRUE.

use.cached.mult
logical Flag telling whether multiplier values should be cached between calls.

use.hinges
logical Flag indicating whether to insert "hinges" into the spectral data before integration so as to reduce interpolation errors at the boundaries of the wavebands.

Value

A single numeric value giving the ratio moles-photons per Joule.

Note

The default for the w.band parameter is a waveband covering the whole range of w.length.

See also

Examples

# photons:energy ratio
with(sun.data, photons_energy_ratio(w.length, s.e.irrad, new_waveband(400,500)))

#> [1] 3.778834e-06

# photons:energy ratio for whole spectrum
with(sun.data, photons_energy_ratio(w.length, s.e.irrad))

#> [1] 4.66451e-06