Print a spectral object

Print method for objects of spectral classes.

# S3 method for generic_spct
print(x, ..., n = NULL, width = NULL)

# S3 method for generic_mspct
print(x, ..., n = NULL, width = NULL,
  n.members = 10)

Arguments

x
An object of one of the summary classes for spectra

...
not used in current version

n
Number of rows to show. If NULL, the default, will print all rows if less than option dplyr.print_max. Otherwise, will print dplyr.print_min

width
Width of text output to generate. This defaults to NULL, which means use getOption("width") and only display the columns that fit on one screen. You can also set option(dplyr.width = Inf) to override this default and always print all columns.

n.members
numeric Number of members of the collection to print.

Value

Returns x invisibly.

Methods (by class)

generic_mspct:

Note

At the moment just a modified copy of dplyr:::print.tbl_df.

Examples

print(sun.spct)

#> Object: source_spct [522 x 3]
#> Wavelength range 280 to 800 nm, step 0.9230769 to 1 nm 
#> Label: sunlight, simulated 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647355 E 
#> Time unit 1s
#> 
#> # A tibble: 522 x 3
#>    w.length s.e.irrad s.q.irrad
#>       <dbl>     <dbl>     <dbl>
#>  1     280.        0.        0.
#>  2     281.        0.        0.
#>  3     282.        0.        0.
#>  4     283.        0.        0.
#>  5     284.        0.        0.
#>  6     285.        0.        0.
#>  7     286.        0.        0.
#>  8     286.        0.        0.
#>  9     287.        0.        0.
#> 10     288.        0.        0.
#> # ... with 512 more rows

print(sun.spct, n = 5)

#> Object: source_spct [522 x 3]
#> Wavelength range 280 to 800 nm, step 0.9230769 to 1 nm 
#> Label: sunlight, simulated 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647355 E 
#> Time unit 1s
#> 
#> # A tibble: 522 x 3
#>   w.length s.e.irrad s.q.irrad
#>      <dbl>     <dbl>     <dbl>
#> 1     280.        0.        0.
#> 2     281.        0.        0.
#> 3     282.        0.        0.
#> 4     283.        0.        0.
#> 5     284.        0.        0.
#> # ... with 517 more rows