Reflectance

Function to calculate the mean, total, or other summary of reflectance for spectral data stored in a reflector_spct or in an object_spct.

reflectance(spct, w.band, quantity, wb.trim, use.hinges, ...)

# S3 method for default
reflectance(spct, w.band, quantity, wb.trim, use.hinges, ...)

# S3 method for reflector_spct
reflectance(spct, w.band = NULL,
  quantity = "average", wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.hinges = getOption("photobiology.use.hinges", default =
  NULL), ...)

# S3 method for object_spct
reflectance(spct, w.band = NULL, quantity = "average",
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.hinges = getOption("photobiology.use.hinges", default = NULL), ...)

# S3 method for reflector_mspct
reflectance(spct, w.band = NULL,
  quantity = "average", wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.hinges = getOption("photobiology.use.hinges", default =
  NULL), ..., attr2tb = NULL, idx = !is.null(names(spct)),
  .parallel = FALSE, .paropts = NULL)

# S3 method for object_mspct
reflectance(spct, w.band = NULL,
  quantity = "average", wb.trim = getOption("photobiology.waveband.trim",
  default = TRUE), use.hinges = getOption("photobiology.use.hinges", default =
  NULL), ..., attr2tb = NULL, idx = !is.null(names(spct)),
  .parallel = FALSE, .paropts = NULL)

Arguments

spct
an R object

w.band
waveband or list of waveband objects or a numeric vector of length two. The waveband(s) determine the region(s) of the spectrum that are summarized. If a numeric range is supplied a waveband object is constructed on the fly from it.

quantity
character string One of "total", "average" or "mean", "contribution", "contribution.pc", "relative" or "relative.pc"

wb.trim
logical Flag telling if wavebands crossing spectral data boundaries are trimmed or ignored

use.hinges
logical Flag indicating whether to insert "hinges" into the spectral data before integration so as to reduce interpolation errors at the boundaries of the wavebands.

...
other arguments

attr2tb
character vector, see add_attr2tb for the syntax for attr2tb passed as is to formal parameter col.names.

idx
logical whether to add a column with the names of the elements of spct

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

Value

A named numeric vector in the case of methods for individual spectra, with one value for each waveband passed to parameter w.band. A data.frame in the case of collections of spectra, containing one column for each waveband object, an index column with the names of the spectra, and optionally additional columns with metadata values retrieved from the attributes of the member spectra.

By default values are only integrated, but depending on the argument passed to parameter quantity they can be re-expressed as relative fractions or percentages. In the case of vector output, names attribute is set to the name of the corresponding waveband unless a named list is supplied in which case the names of the list members are used.

Methods (by class)

default: Default for generic function

reflector_spct: Specialization for reflector_spct

object_spct: Specialization for object_spct

reflector_mspct: Calculates reflectance from a reflector_mspct

object_mspct: Calculates reflectance from a object_mspct

Note

The use.hinges parameter controls speed optimization. The defaults should be suitable in most cases. Only the range of wavelengths in the wavebands is used and all BSWFs are ignored.

Examples

reflectance(black_body.spct, waveband(c(400,700)))

#>  range.400.700 
#>              0 
#> attr(,"Rfr.type")
#> [1] "total"
#> attr(,"radiation.unit")
#> [1] "reflectance average"

reflectance(white_body.spct, waveband(c(400,700)))

#>  range.400.700 
#>              1 
#> attr(,"Rfr.type")
#> [1] "total"
#> attr(,"radiation.unit")
#> [1] "reflectance average"