Relative Air Mass (AM)

Approximate relative air mass (AM) from sun elevation or sun zenith angle.

relative_AM(elevation.angle = NULL, zenith.angle = NULL,
  occluded.value = NA)

Arguments

elevation.angle, zenith.angle
numeric vector Angle in degrees for the sun position. An argument should be passed to one and only one of elevation_angle or zenith_angle.

occluded.value
numeric Value to return when elevation angle is negative (sun below the horizon).

Details

This is an implementation of equation (3) in Kasten and Young (1989). This equation is only an approximation to the tabulated values in the same paper. Returned values are rounded to three significant digits.

Note

Although relative air mass is not defined when the sun is not visible, returning a value different from the default NA might be useful in some cases.

References

F. Kasten, A. T. Young (1989) Revised optical air mass tables and approximation formula. Applied Optics, 28, 4735-. doi:10.1364/ao.28.004735.

Examples

relative_AM(c(90, 60, 30, 1, -10))

#> [1]  0.999  1.150  1.990 26.300     NA

relative_AM(c(90, 60, 30, 1, -10), occluded.value = Inf)

#> [1]  0.999  1.150  1.990 26.300    Inf

relative_AM(zenith.angle = 0)

#> [1] 0.999