Integrated response

Calculate average photon- or energy-based photo-response.

response(spct, w.band, unit.out, quantity, time.unit, wb.trim, use.hinges, ...)

# S3 method for default
response(spct, w.band, unit.out, quantity, time.unit, wb.trim,
  use.hinges, ...)

# S3 method for response_spct
response(spct, w.band = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  quantity = "total", time.unit = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.hinges = getOption("photobiology.use.hinges", default = NULL), ...)

# S3 method for response_mspct
response(spct, w.band = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  quantity = "total", time.unit = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.hinges = getOption("photobiology.use.hinges", default = NULL), ...,
  idx = !is.null(names(spct)))

Arguments

spct
an R object of class "generic_spct"

w.band
waveband or list of waveband objects or a numeric vector of length two. The waveband(s) determine the region(s) of the spectrum that are summarized. If a numeric range is supplied a waveband object is constructed on the fly from it.

unit.out
character Allowed values "energy", and "photon", or its alias "quantum"

quantity
character string One of "total", "average" or "mean", "contribution", "contribution.pc", "relative" or "relative.pc"

time.unit
character or lubridate::duration

wb.trim
logical Flag telling if wavebands crossing spectral data boundaries are trimmed or ignored

use.hinges
logical indicating whether to use hinges to reduce interpolation errors

...
other arguments

idx
logical whether to add a column with the names of the elements of spct

Value

A single numeric value expressed either as a fraction of one or a percentage, or a vector of the same length as the list of wave.bands. The quantity returned depends on the value of quantity. Whether it is expressed in energy-based or photon-based units depends on unit.out.

Methods (by class)

default: Default for generic function

response_spct: Method for response spectra.

response_mspct: Calculates response from a response_mspct

Note

The parameter use.hinges controls speed optimization. The defaults should be suitable in mosts cases. Only the range of wavelengths in the wavebands is used and all BSWFs are ignored.

See also

Other response functions: e_response, q_response