Rounding of Numbers

ceiling takes a single numeric argument x and returns a numeric vector containing the smallest integers not less than the corresponding elements of x. \ floor takes a single numeric argument x and returns a numeric vector containing the largest integers not greater than the corresponding elements of x. \ trunc takes a single numeric argument x and returns a numeric vector containing the integers formed by truncating the values in x toward 0. \ round rounds the values in its first argument to the specified number of decimal places (default 0). \ signif rounds the values in its first argument to the specified number of significant digits. \ The functions are applied to the spectral data, not the wavelengths. The quantity in the spectrum to which the function is applied depends on the class of x and the current value of output options.

# S3 method for generic_spct
round(x, digits = 0)

# S3 method for generic_spct
signif(x, digits = 6)

# S3 method for generic_spct
ceiling(x)

# S3 method for generic_spct
floor(x)

# S3 method for generic_spct
trunc(x, ...)

Arguments

x
an object of class "generic_spct" or a derived class.

digits
integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed (see ‘Details’).

...
arguments to be passed to methods.

See also

Other math operators and functions: MathFun, ^.generic_spct, convolve_each, div-.generic_spct, log, minus-.generic_spct, mod-.generic_spct, plus-.generic_spct, sign, slash-.generic_spct, times-.generic_spct