Set the "Rfr.type" attribute

Function to set by reference the "Rfr.type" attribute of an existing reflector_spct or object_spct object.

setRfrType(x, Rfr.type = c("total", "specular"))

Arguments

x
a reflector_spct or an object_spct object

Rfr.type
a character string, either "total" or "specular"

Value

x

Note

This function alters x itself by reference and in addition returns x invisibly. If x is not a reflector_spct or object_spct object, x is not modified. The behaviour of this function is 'unusual' in that the default for parameter Rfr.type is used only if x does not already have this attribute set.