Convert a 'wide' or untidy data frame into a collection of spectra

Convert a data frame object into a "multi spectrum" object by constructing a an object of a multi-spct class, converting numeric columns other than wavelength into individual spct objects.

split2mspct(x, member.class = NULL, spct.data.var = NULL,
  w.length.var = "w.length", idx.var = NULL, ncol = 1, byrow = FALSE,
  ...)

split2source_mspct(x, spct.data.var = "s.e.irrad",
  w.length.var = "w.length", idx.var = NULL, ncol = 1, byrow = FALSE,
  ...)

split2response_mspct(x, spct.data.var = "s.e.response",
  w.length.var = "w.length", idx.var = NULL, ncol = 1, byrow = FALSE,
  ...)

split2filter_mspct(x, spct.data.var = "Tfr", w.length.var = "w.length",
  idx.var = NULL, ncol = 1, byrow = FALSE, ...)

split2reflector_mspct(x, spct.data.var = "Rfr", w.length.var = "w.length",
  idx.var = NULL, ncol = 1, byrow = FALSE, ...)

split2cps_mspct(x, spct.data.var = "cps", w.length.var = "w.length",
  idx.var = NULL, ncol = 1, byrow = FALSE, ...)

split2raw_mspct(x, spct.data.var = "count", w.length.var = "w.length",
  idx.var = NULL, ncol = 1, byrow = FALSE, ...)

Arguments

x
data frame

member.class
character Class of the collection members

spct.data.var
character Name of the spctral data argument in the object constructor for member.class

w.length.var
character Name of column containing wavelength data in nanometres

idx.var
character Name of column containing data to be copied unchanged to each spct object

ncol
integer Number of 'virtual' columns in data

byrow
logical If ncol > 1 how to read in the data

...
additional named arguments passed to the member constructor function.

See also

Other collections of spectra classes family: generic_mspct, subset2mspct