Stepsize

Function that returns the range of step sizes in an object. Range of differences between successive sorted values.

wl_stepsize(x, ...)

stepsize(x, ...)

# S3 method for default
stepsize(x, ...)

# S3 method for numeric
stepsize(x, ...)

# S3 method for generic_spct
stepsize(x, ...)

# S3 method for generic_mspct
stepsize(x, ..., idx = !is.null(names(x)))

Arguments

x
an R object

...
not used in current version

idx
logical whether to add a column with the names of the elements of spct

Value

A numeric vector of length 2 with min and maximum stepsize values.

Methods (by class)

default: Default function usable on numeric vectors.

numeric: Method for numeric vectors.

generic_spct: Method for "generic_spct" objects.

generic_mspct: Method for "generic_mspct" objects.

See also

Examples

stepsize(sun.spct)

#> [1] 0.9230769 1.0000000

wl_stepsize(sun.spct)

#> [1] 0.9230769 1.0000000

stepsize(sun.spct)

#> [1] 0.9230769 1.0000000