Daily solar spectral irradiance (simulated)

A dataset containing the wavelengths at a 1 nm interval and the corresponding spectral (energy) irradiance. Values simulated for 2 June 2012, at Helsinki, under clear sky conditions. The variables are as follows:

sun.daily.spct

Format

A source_spct object with 511 rows and 3 variables

Details

w.length (nm), range 290 to 800 nm.

s.e.irrad (J d-1 m-2 nm-1)

s.q.irrad (mol d-1 m-2 nm-1)

Note

The simulations are based on libRadTran using hourly mean global radiation measurements to estimate cloud cover. The simulations were for each hour and the results integrated for the whole day.

References

Lindfors, A.; Heikkilä, A.; Kaurola, J.; Koskela, T. & Lakkala, K. (2009) Reconstruction of Solar Spectral Surface UV Irradiances Using Radiative Transfer Simulations. Photochemistry and Photobiology, 85: 1233–1239

See also

Examples

sun.daily.spct

#> Object: source_spct [522 x 3]
#> Wavelength range 280 to 800 nm, step 0.9 to 1 nm 
#> Label: sunlight, simulated 
#> Measured on 2012-06-01 UTC 
#> Measured at 60.2091078 N, 24.9647355 E 
#> Time unit 86400s (~1 days)
#> 
#> # A tibble: 522 x 3
#>    w.length s.e.irrad s.q.irrad
#>       <dbl>     <dbl>     <dbl>
#>  1     280.        0.        0.
#>  2     281.        0.        0.
#>  3     282.        0.        0.
#>  4     283.        0.        0.
#>  5     284.        0.        0.
#>  6     284.        0.        0.
#>  7     285.        0.        0.
#>  8     286.        0.        0.
#>  9     287.        0.        0.
#> 10     288.        0.        0.
#> # ... with 512 more rows