Solar angles

This function returns the solar angles for a given time and location.

sun_angles(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE, use.fast.code = TRUE)

sun_angles_fast(time, tz, geocode, use.refraction)

sun_elevation(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

sun_zenith_angle(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

sun_azimuth(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

Arguments

time
A "vector" of POSIXct Time, with any valid time zone (TZ) is allowed, default is current time.

tz
character string indicating time zone to be used in output, if a vector of length 2, first value is used for sunrise and second for sunset.

geocode
data frame with variables lon and lat as numeric values (degrees), nrow > 1, allowed.

use.refraction
logical Flag indicating whether to correct for fraction in the atmosphere.

use.fast.code
logical Can be used to disable optimizations, for testing.

Value

A data.frame with variables time (in same TZ as input), TZ, solartime, longitude, latitude, address, azimuth, and elevation.

Note

The special, and very frequent case of many time points for a single geocode and time zone is heavily optimized for performance. If the arguments passed fullfil this condition, optimized code is used.

See also

Other astronomy related functions: day_night, format.solar_time, is.solar_time, print.solar_time, solar_time

Examples

library(lubridate)
sun_angles()

#>                  time tz solartime longitude latitude   address azimuth
#> 1 2017-06-18 12:47:52     09:46:38         0     51.5 Greenwich 124.957
#>   elevation
#> 1  52.01922

sun_angles(ymd_hms("2014-09-23 12:00:00"))

#>                  time  tz solartime longitude latitude   address  azimuth
#> 1 2014-09-23 12:00:00 UTC  12:07:41         0     51.5 Greenwich 182.4467
#>   elevation
#> 1  38.20657

sun_angles(ymd_hms("2014-09-23 12:00:00"),
           geocode = data.frame(lat=60, lon=0))

#>                  time  tz solartime longitude latitude address  azimuth
#> 1 2014-09-23 12:00:00 UTC  12:07:41         0       60    <NA> 182.2136
#>   elevation
#> 1  29.71356

sun_angles(ymd_hms("2014-09-23 12:00:00") + minutes((0:6) * 10))

#>                  time  tz solartime longitude latitude   address  azimuth
#> 1 2014-09-23 12:00:00 UTC  12:07:41         0     51.5 Greenwich 182.4467
#> 2 2014-09-23 12:10:00 UTC  12:17:41         0     51.5 Greenwich 185.6235
#> 3 2014-09-23 12:20:00 UTC  12:27:41         0     51.5 Greenwich 188.7868
#> 4 2014-09-23 12:30:00 UTC  12:37:41         0     51.5 Greenwich 191.9294
#> 5 2014-09-23 12:40:00 UTC  12:47:41         0     51.5 Greenwich 195.0447
#> 6 2014-09-23 12:50:00 UTC  12:57:42         0     51.5 Greenwich 198.1265
#> 7 2014-09-23 13:00:00 UTC  13:07:42         0     51.5 Greenwich 201.1693
#>   elevation
#> 1  38.20657
#> 2  38.09432
#> 3  37.89637
#> 4  37.61378
#> 5  37.24803
#> 6  36.80100
#> 7  36.27491