Tag a spectrum

Spectra are tagged by adding variables and attributes containing color definitions, labels, and a factor following the wavebands given in w.band.

tag(x, ...)

# S3 method for default
tag(x, ...)

# S3 method for generic_spct
tag(x, w.band = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.hinges = TRUE, short.names = TRUE, byref = FALSE, ...)

# S3 method for generic_mspct
tag(x, w.band = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.hinges = TRUE, short.names = TRUE, byref = FALSE, ...)

Arguments

x
an R object

...
not used in current version

w.band
waveband or list of waveband objects The waveband(s) determine the region(s) of the spectrum that are tagged

wb.trim
logical Flag telling if wavebands crossing spectral data boundaries are trimmed or ignored

use.hinges
logical Flag indicating whether to use hinges to reduce interpolation errors

short.names
logical Flag indicating whether to use short or long names for wavebands

byref
logical Flag indicating if new object will be created by reference or by copy of x

Methods (by class)

default: Default method for generic

generic_spct: Tag one of generic_spct, and derived classes including source_spct, filter_spct, reflector_spct, object_spct, and response_spct.

generic_mspct: Tag one of generic_mspct, and derived classes including source_mspct, filter_mspct, reflector_mspct, object_mspct, and response_mspct.

Note

NULL as w.band argument does not add any new tags, instead it removes existing tags if present. NA, the default, as w.band argument removes existing waveband tags if present and sets the wl.color variable. If a waveband object or a list of wavebands is supplied as argument then tagging is based on them, and wl.color is also set.

See also

Other tagging and related functions: is_tagged, untag, wb2rect_spct, wb2spct, wb2tagged_spct

Examples

tag(sun.spct)

#> Object: source_spct [523 x 6]
#> Wavelength range 280 to 800 nm, step 1.023182e-12 to 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 523 x 6
#>    w.length s.e.irrad s.q.irrad wl.color wb.color   wb.f
#>       <dbl>     <dbl>     <dbl>    <chr>    <chr> <fctr>
#>  1 280.0000         0         0  #000000  #554340  Total
#>  2 280.9231         0         0  #000000  #554340  Total
#>  3 281.8462         0         0  #000000  #554340  Total
#>  4 282.7692         0         0  #000000  #554340  Total
#>  5 283.6923         0         0  #000000  #554340  Total
#>  6 284.6154         0         0  #000000  #554340  Total
#>  7 285.5385         0         0  #000000  #554340  Total
#>  8 286.4615         0         0  #000000  #554340  Total
#>  9 287.3846         0         0  #000000  #554340  Total
#> 10 288.3077         0         0  #000000  #554340  Total
#> # ... with 513 more rows

tag(sun.spct, list(A = waveband(c(300,3005))))

#> Object: source_spct [524 x 6]
#> Wavelength range 280 to 800 nm, step 1.023182e-12 to 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 524 x 6
#>    w.length s.e.irrad s.q.irrad wl.color wb.color   wb.f
#>       <dbl>     <dbl>     <dbl>    <chr>    <chr> <fctr>
#>  1 280.0000         0         0  #000000     <NA>   <NA>
#>  2 280.9231         0         0  #000000     <NA>   <NA>
#>  3 281.8462         0         0  #000000     <NA>   <NA>
#>  4 282.7692         0         0  #000000     <NA>   <NA>
#>  5 283.6923         0         0  #000000     <NA>   <NA>
#>  6 284.6154         0         0  #000000     <NA>   <NA>
#>  7 285.5385         0         0  #000000     <NA>   <NA>
#>  8 286.4615         0         0  #000000     <NA>   <NA>
#>  9 287.3846         0         0  #000000     <NA>   <NA>
#> 10 288.3077         0         0  #000000     <NA>   <NA>
#> # ... with 514 more rows