Trim (or expand) head and/or tail

Trim tails of a spectrum based on wavelength limits, interpolating the values at the boundaries.Trimming is needed for example to remove short wavelength noise when the measured spectrum extends beyond the known emission spectrum of the measured light source. Occasionally one may want also to expand the wavelength range.

trim_tails(x, y, low.limit = min(x), high.limit = max(x),
  use.hinges = TRUE, fill = NULL, verbose = TRUE)

Arguments

x
numeric vector of wavelengths.

y
numeric vector of values for a spectral quantity.

low.limit
smallest x-value to be kept (defaults to smallest x-value in input).

high.limit
largest x-value to be kept (defaults to largest x-value in input).

use.hinges
logical Flag indicating whether to insert "hinges" into the spectral data before integration so as to reduce interpolation errors at the boundaries of the wavebands.

fill
if fill == NULL then tails are deleted, otherwise tails of y are filled with the value of fill.

verbose
logical Use to suppress warnings.

Value

A data.frame with variables x and y.

Note

When expanding an spectrum, if fill == NULL, expansion is not performed with a warning.

See also

Other low-level functions operating on numeric vectors.: as_energy, as_quantum_mol, calc_multipliers, div_spectra, energy_irradiance, energy_ratio, insert_hinges, integrate_xy, interpolate_spectrum, irradiance, l_insert_hinges, oper_spectra, photon_irradiance, photon_ratio, photons_energy_ratio, prod_spectra, s_e_irrad2rgb, split_energy_irradiance, split_photon_irradiance, subt_spectra, sum_spectra, v_insert_hinges

Examples

head(sun.data)

#> # A tibble: 6 x 3
#>   w.length  s.e.irrad s.q.irrad
#>      <dbl>      <dbl>     <dbl>
#> 1     293. 0.00000261  6.39e-12
#> 2     294. 0.00000614  1.51e-11
#> 3     295. 0.0000218   5.37e-11
#> 4     296. 0.0000678   1.68e-10
#> 5     297. 0.000153    3.81e-10
#> 6     298. 0.000367    9.14e-10

head(with(sun.data,
     trim_tails(w.length, s.e.irrad, low.limit=300)))

#>      x           y
#> 8  300 0.001264554
#> 9  301 0.002623718
#> 10 302 0.003922583
#> 11 303 0.008974134
#> 12 304 0.011655666
#> 13 305 0.017991275

head(with(sun.data,
     trim_tails(w.length, s.e.irrad, low.limit=300, fill=NULL)))

#>      x           y
#> 8  300 0.001264554
#> 9  301 0.002623718
#> 10 302 0.003922583
#> 11 303 0.008974134
#> 12 304 0.011655666
#> 13 305 0.017991275