Waveband constructor method

Constructor for "waveband" objects that can be used as input when calculating irradiances.

waveband(x = NULL, weight = NULL, SWF.e.fun = NULL, SWF.q.fun = NULL,
  norm = NULL, SWF.norm = NULL, hinges = NULL, wb.name = NULL,
  wb.label = wb.name)

new_waveband(w.low, w.high, weight = NULL, SWF.e.fun = NULL,
  SWF.q.fun = NULL, norm = NULL, SWF.norm = NULL, hinges = NULL,
  wb.name = NULL, wb.label = wb.name)

Arguments

x
any R object on which applying the function range yields an vector of two numeric values, describing a range of wavelengths (nm)

weight
a character string "SWF" or "BSWF", use NULL (the default) to indicate no weighting used when calculating irradiance

SWF.e.fun
a function giving multipliers for a spectral weighting function (energy) as a function of wavelength (nm)

SWF.q.fun
a function giving multipliers for a spectral weighting function (quantum) as a function of wavelength (nm)

norm
a single numeric value indicating the wavelength at which the SWF should be normalized to 1.0, in nm. "NULL" means no normalization.

SWF.norm
a numeric value giving the native normalization wavelength (nm) used by SWF.e.fun and SWF.q.fun

hinges
a numeric vector giving the wavelengths at which the s.irrad should be inserted by interpolation, no interpolation is indicated by an empty vector (numeric(0)), if NULL then interpolation will take place at both ends of the band.

wb.name
character string giving the name for the waveband defined, default is NULL

wb.label
character string giving the label of the waveband to be used for plotting, default is wb.name

w.low
numeric value, wavelength at the short end of the band (nm)

w.high
numeric value, wavelength at the long end of the band (nm)

Value

a waveband object

Functions

new_waveband: A less flexible variant

See also

Other waveband constructors: split_bands

Examples

waveband(c(400,700))

#> range.400.700 
#> low (nm) 400 
#> high (nm) 700 
#> weighted none 

new_waveband(400,700)

#> range.400.700 
#> low (nm) 400 
#> high (nm) 700 
#> weighted none