Calculate energy to quantum multipliers

Gives multipliers as a function of wavelength, for converting from energy to photon (quantum) units (number of photons as default, or moles of photons).

e2quantum_multipliers(w.length, molar = FALSE)

Arguments

w.length
numeric Vector of wavelengths (nm)

molar
logical Flag indicating whether output should be in moles or numbers

Value

A numeric array of multipliers

See also

Other quantity conversion functions: A2T, T2A, as_energy, as_quantum_mol, as_quantum, e2qmol_multipliers, e2q, q2e

Examples

with(sun.data, e2quantum_multipliers(w.length))

#>   [1] 1.474975e+18 1.480009e+18 1.485043e+18 1.490077e+18 1.495111e+18
#>   [6] 1.500145e+18 1.505179e+18 1.510213e+18 1.515247e+18 1.520281e+18
#>  [11] 1.525315e+18 1.530349e+18 1.535383e+18 1.540417e+18 1.545451e+18
#>  [16] 1.550485e+18 1.555519e+18 1.560553e+18 1.565587e+18 1.570621e+18
#>  [21] 1.575656e+18 1.580690e+18 1.585724e+18 1.590758e+18 1.595792e+18
#>  [26] 1.600826e+18 1.605860e+18 1.610894e+18 1.615928e+18 1.620962e+18
#>  [31] 1.625996e+18 1.631030e+18 1.636064e+18 1.641098e+18 1.646132e+18
#>  [36] 1.651166e+18 1.656200e+18 1.661234e+18 1.666268e+18 1.671302e+18
#>  [41] 1.676336e+18 1.681370e+18 1.686404e+18 1.691439e+18 1.696473e+18
#>  [46] 1.701507e+18 1.706541e+18 1.711575e+18 1.716609e+18 1.721643e+18
#>  [51] 1.726677e+18 1.731711e+18 1.736745e+18 1.741779e+18 1.746813e+18
#>  [56] 1.751847e+18 1.756881e+18 1.761915e+18 1.766949e+18 1.771983e+18
#>  [61] 1.777017e+18 1.782051e+18 1.787085e+18 1.792119e+18 1.797153e+18
#>  [66] 1.802187e+18 1.807222e+18 1.812256e+18 1.817290e+18 1.822324e+18
#>  [71] 1.827358e+18 1.832392e+18 1.837426e+18 1.842460e+18 1.847494e+18
#>  [76] 1.852528e+18 1.857562e+18 1.862596e+18 1.867630e+18 1.872664e+18
#>  [81] 1.877698e+18 1.882732e+18 1.887766e+18 1.892800e+18 1.897834e+18
#>  [86] 1.902868e+18 1.907902e+18 1.912936e+18 1.917970e+18 1.923005e+18
#>  [91] 1.928039e+18 1.933073e+18 1.938107e+18 1.943141e+18 1.948175e+18
#>  [96] 1.953209e+18 1.958243e+18 1.963277e+18 1.968311e+18 1.973345e+18
#>  [ reached getOption("max.print") -- omitted 408 entries ]

with(sun.data, e2quantum_multipliers(w.length, molar = TRUE))

#>   [1] 2.449253e-06 2.457612e-06 2.465971e-06 2.474331e-06 2.482690e-06
#>   [6] 2.491049e-06 2.499408e-06 2.507767e-06 2.516127e-06 2.524486e-06
#>  [11] 2.532845e-06 2.541204e-06 2.549564e-06 2.557923e-06 2.566282e-06
#>  [16] 2.574641e-06 2.583000e-06 2.591360e-06 2.599719e-06 2.608078e-06
#>  [21] 2.616437e-06 2.624797e-06 2.633156e-06 2.641515e-06 2.649874e-06
#>  [26] 2.658233e-06 2.666593e-06 2.674952e-06 2.683311e-06 2.691670e-06
#>  [31] 2.700030e-06 2.708389e-06 2.716748e-06 2.725107e-06 2.733466e-06
#>  [36] 2.741826e-06 2.750185e-06 2.758544e-06 2.766903e-06 2.775263e-06
#>  [41] 2.783622e-06 2.791981e-06 2.800340e-06 2.808699e-06 2.817059e-06
#>  [46] 2.825418e-06 2.833777e-06 2.842136e-06 2.850496e-06 2.858855e-06
#>  [51] 2.867214e-06 2.875573e-06 2.883933e-06 2.892292e-06 2.900651e-06
#>  [56] 2.909010e-06 2.917369e-06 2.925729e-06 2.934088e-06 2.942447e-06
#>  [61] 2.950806e-06 2.959166e-06 2.967525e-06 2.975884e-06 2.984243e-06
#>  [66] 2.992602e-06 3.000962e-06 3.009321e-06 3.017680e-06 3.026039e-06
#>  [71] 3.034399e-06 3.042758e-06 3.051117e-06 3.059476e-06 3.067835e-06
#>  [76] 3.076195e-06 3.084554e-06 3.092913e-06 3.101272e-06 3.109632e-06
#>  [81] 3.117991e-06 3.126350e-06 3.134709e-06 3.143068e-06 3.151428e-06
#>  [86] 3.159787e-06 3.168146e-06 3.176505e-06 3.184865e-06 3.193224e-06
#>  [91] 3.201583e-06 3.209942e-06 3.218302e-06 3.226661e-06 3.235020e-06
#>  [96] 3.243379e-06 3.251738e-06 3.260098e-06 3.268457e-06 3.276816e-06
#>  [ reached getOption("max.print") -- omitted 408 entries ]