Peaks or local maxima

Function that returns a subset of an R object with observations corresponding to local maxima.

peaks(x, span, ignore_threshold, strict, ...)

# S3 method for default
peaks(x, span, ignore_threshold, strict, ...)

# S3 method for numeric
peaks(x, span = 5, ignore_threshold, strict = TRUE, ...)

# S3 method for generic_spct
peaks(x, span, ignore_threshold, strict, ...)

# S3 method for source_spct
peaks(x, span = 5, ignore_threshold = 0,
  strict = TRUE, unit.out = getOption("photobiology.radiation.unit", default
  = "energy"), ...)

# S3 method for response_spct
peaks(x, span = 5, ignore_threshold = 0,
  strict = TRUE, unit.out = getOption("photobiology.radiation.unit", default
  = "energy"), ...)

# S3 method for filter_spct
peaks(x, span = 5, ignore_threshold = 0,
  strict = TRUE, filter.qty = getOption("photobiology.filter.qty", default =
  "transmittance"), ...)

# S3 method for reflector_spct
peaks(x, span = 5, ignore_threshold = 0,
  strict = TRUE, ...)

# S3 method for cps_spct
peaks(x, span = 5, ignore_threshold = 0, strict = TRUE,
  ...)

# S3 method for generic_mspct
peaks(x, span = 5, ignore_threshold = 0,
  strict = TRUE, ..., .parallel = FALSE, .paropts = NULL)

Arguments

x
an R object

span
a peak is defined as an element in a sequence which is greater than all other elements within a window of width span centered at that element. The default value is 3, meaning that a peak is bigger than both of its neighbors. Default: 3.

ignore_threshold
numeric value between 0.0 and 1.0 indicating the relative size compared to tallest peak threshold below which peaks will be ignored.

strict
logical flag: if TRUE, an element must be strictly greater than all other values in its window to be considered a peak. Default: TRUE.

...
ignored

unit.out
character One of "energy" or "photon"

filter.qty
character One of "transmittance" or "absorbance"

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

Value

A subset of x with rows corresponding to local maxima.

Methods (by class)

default: Default function usable on numeric vectors.

numeric: Default function usable on numeric vectors.

generic_spct: Method for "generic_spct" objects.

source_spct: Method for "source_spct" objects.

response_spct: Method for "response_spct" objects.

filter_spct: Method for "filter_spct" objects.

reflector_spct: Method for "reflector_spct" objects.

cps_spct: Method for "cps_spct" objects.

generic_mspct: Method for "cps_spct" objects.

See also

Examples

peaks(sun.spct)

#> Object: source_spct [76 x 2]
#> Wavelength range 321 to 798 nm, step 3 to 20 nm 
#> Label: sunlight, simulated 
#> Measured on 2010-06-22 09:51:00 UTC 
#> Measured at 60.2091078 N, 24.9647355 E 
#> Time unit 1s
#> 
#> # A tibble: 76 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1     321.     0.182
#>  2     330.     0.330
#>  3     335.     0.313
#>  4     340.     0.335
#>  5     343.     0.338
#>  6     347.     0.321
#>  7     350.     0.345
#>  8     354.     0.376
#>  9     360.     0.371
#> 10     366.     0.449
#> # ... with 66 more rows