Solar angles

This function returns the solar angles for a given time and location.

sun_angles(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

sun_elevation(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

sun_zenith_angle(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

sun_azimuth(time = lubridate::now(), tz = lubridate::tz(time),
  geocode = data.frame(lon = 0, lat = 51.5, address = "Greenwich"),
  use.refraction = FALSE)

Arguments

time
POSIXct Time, any valid time zone (TZ) is allowed, default is current time

tz
character string indicating time zone to be used in output.

geocode
data frame with variables lon and lat as numeric values (degrees).

use.refraction
logical Flag indicating whether to correct for fraction in the atmosphere

Value

A data.frame with variables time (in same TZ as input), TZ, solartime, longitude, latitude, address, azimuth, and elevation.

See also

Other astronomy related functions: day_night, format.solar_time, is.solar_time, print.solar_time, solar_time

Examples

library(lubridate)
sun_angles()

#>                  time tz solartime longitude latitude   address  azimuth
#> 1 2017-04-29 01:28:56     22:31:34         0     51.5 Greenwich 336.9964
#>   elevation
#> 1 -21.22489

sun_angles(ymd_hms("2014-09-23 12:00:00"))

#>                  time  tz solartime longitude latitude   address  azimuth
#> 1 2014-09-23 12:00:00 UTC  12:07:41         0     51.5 Greenwich 182.4467
#>   elevation
#> 1  38.20657

sun_angles(ymd_hms("2014-09-23 12:00:00"),
           geocode = data.frame(lat=60, lon=0))

#>                  time  tz solartime longitude latitude address  azimuth
#> 1 2014-09-23 12:00:00 UTC  12:07:41         0       60    <NA> 182.2136
#>   elevation
#> 1  29.71356